refactor: updated imports and scopes in all components
