+++ docs/components/feedback/star-banner.tsx
@@ -2,7 +2,7 @@ import { Button, Center, HStack, Text, forwardRef } from "@yamada-ui/react"
 import type { CenterProps } from "@yamada-ui/react"
 import { Star } from "components/media-and-icons"
 import { CONSTANT } from "constant"
-import { useI18n } from "contexts/i18n-context"
+import { useI18n } from "contexts"
 import { memo } from "react"
 
 export type StarBannerProps = CenterProps & {}
