+++ docs/components/forms/search.tsx
@@ -1,3 +1,4 @@
+import { File, Hash, Search as SearchIcon } from "@yamada-ui/lucide"
 import {
   ui,
   HStack,
@@ -17,10 +18,9 @@ import {
   useUpdateEffect,
   IconButton,
 } from "@yamada-ui/react"
-import { File, Hash, Search as SearchIcon } from "@yamada-ui/lucide"
 import type { StackProps, ModalProps, ButtonProps } from "@yamada-ui/react"
-import { useI18n } from "contexts/i18n-context"
-import { useEventListener } from "hooks/use-event-listener"
+import { useI18n } from "contexts"
+import { useEventListener } from "hooks"
 import { matchSorter } from "match-sorter"
 import NextLink from "next/link"
 import { useRouter } from "next/router"
