+++ docs/components/layouts/footer.tsx
@@ -9,7 +9,7 @@ import {
 import type { CenterProps } from "@yamada-ui/react"
 import { Github, X } from "components/media-and-icons"
 import { CONSTANT } from "constant"
-import { useI18n } from "contexts/i18n-context"
+import { useI18n } from "contexts"
 import { memo } from "react"
 
 export type FooterProps = CenterProps & {}
