+++ docs/components/layouts/header.tsx
@@ -54,12 +54,11 @@ import { Search, SearchButton } from "components/forms"
 import { Discord, Github } from "components/media-and-icons"
 import { NextLinkIconButton, Tree } from "components/navigation"
 import { CONSTANT } from "constant"
-import { useI18n } from "contexts/i18n-context"
-import { usePage } from "contexts/page-context"
+import { useI18n, usePage } from "contexts"
 import Link from "next/link"
 import { useRouter } from "next/router"
-import { memo, useEffect, useRef, useState } from "react"
 import type { FC } from "react"
+import { memo, useEffect, useRef, useState } from "react"
 
 export type HeaderProps = CenterProps & {}
