+++ docs/components/mdx/color-mode-button.tsx
@@ -1,6 +1,6 @@
 import { Button, useColorMode, useColorModeValue } from "@yamada-ui/react"
 import type { ButtonProps } from "@yamada-ui/react"
-import { useI18n } from "contexts/i18n-context"
+import { useI18n } from "contexts"
 import type { FC } from "react"
 
 export type ColorModeButtonProps = ButtonProps
