+++ docs/components/mdx/link.tsx
@@ -1,5 +1,5 @@
-import { Text } from "@yamada-ui/react"
 import { ExternalLink } from "@yamada-ui/lucide"
+import { Text } from "@yamada-ui/react"
 import { NextLink } from "components/navigation"
 import type { NextLinkProps } from "components/navigation"
 import type { FC } from "react"
