+++ docs/components/mdx/select-package-managers.tsx
@@ -2,7 +2,7 @@ import { useBreakpoint } from "@yamada-ui/react"
 import { Table } from "@yamada-ui/table"
 import type { Column } from "@yamada-ui/table"
 import { TextWithCodeOrLink } from "components/typography"
-import { useI18n } from "contexts/i18n-context"
+import { useI18n } from "contexts"
 import { useMemo, useState } from "react"
 import type { FC } from "react"
 import { PackageManagers } from "./package-managers"
