+++ docs/components/navigation/link-card.tsx
@@ -10,7 +10,7 @@ import {
 import type { CardProps, StringLiteral } from "@yamada-ui/react"
 import { Label } from "components/data-display"
 import { TextWithCode } from "components/typography"
-import { usePage } from "contexts/page-context"
+import { usePage } from "contexts"
 import Link from "next/link"
 import { memo } from "react"
 import type { ReactNode } from "react"
