+++ docs/components/navigation/related-links.tsx
@@ -2,10 +2,9 @@ import { Button, forwardRef, Link, Wrap } from "@yamada-ui/react"
 import type { ButtonProps, FlexProps, LinkProps } from "@yamada-ui/react"
 import { Github, Npm } from "components/media-and-icons"
 import { CONSTANT } from "constant"
-import { useI18n } from "contexts/i18n-context"
-import { usePage } from "contexts/page-context"
-import { memo } from "react"
+import { useI18n, usePage } from "contexts"
 import type { FC } from "react"
+import { memo } from "react"
 
 export type RelatedLinksProps = FlexProps
