+++ docs/components/navigation/tabs.tsx
@@ -1,6 +1,6 @@
 import { forwardRef, Box, Tabs as UITabs, Tab } from "@yamada-ui/react"
 import type { BoxProps } from "@yamada-ui/react"
-import { usePage } from "contexts/page-context"
+import { usePage } from "contexts"
 import { memo } from "react"
 import { NextLink } from "./next-link"
