+++ docs/contexts/index.ts
@@ -1,0 +1,2 @@
+export * from "./i18n-context"
+export * from "./page-context"
