+++ docs/layouts/index.ts
@@ -1,0 +1,2 @@
+export * from "./top-layout"
+export * from "./document-layout"
