+++ docs/pages/404.page.tsx
@@ -2,9 +2,8 @@ import { Text, VStack } from "@yamada-ui/react"
 import { Section } from "components/layouts"
 import { SEO } from "components/media-and-icons"
 import { NextLinkButton } from "components/navigation"
-import { useI18n } from "contexts/i18n-context"
-import { PageProvider } from "contexts/page-context"
-import { TopLayout } from "layouts/top-layout"
+import { useI18n, PageProvider } from "contexts"
+import { TopLayout } from "layouts"
 import type { InferGetStaticPropsType, NextPage } from "next"
 import { getStaticCommonProps } from "utils/next"
