+++ docs/pages/api/github/webhook/pull-request-review/index.ts
@@ -1,7 +1,7 @@
 import type { PullRequestReviewAction } from "utils/github"
 import type { APIHandler } from "utils/next"
-import { submitted } from "./submitted"
 import { dismissed } from "./dismissed"
+import { submitted } from "./submitted"
 
 export const pullRequestReview: {
   [key in PullRequestReviewAction]?: APIHandler
