+++ docs/pages/examples/[slug].page.tsx
@@ -1,3 +1,4 @@
+import { ArrowRight, Blocks } from "@yamada-ui/lucide"
 import {
   Avatar,
   AvatarGroup,
@@ -12,6 +13,12 @@ import {
   Tooltip,
   VStack,
 } from "@yamada-ui/react"
+import { Section } from "components/layouts"
+import { SEO } from "components/media-and-icons"
+import { NextLinkButton } from "components/navigation"
+import { CONSTANT } from "constant"
+import { useI18n, PageProvider } from "contexts"
+import { TopLayout } from "layouts"
 import type {
   GetStaticPathsContext,
   GetStaticPathsResult,
@@ -21,13 +28,6 @@ import type {
 } from "next"
 import Link from "next/link"
 import { useMemo } from "react"
-import { Section } from "components/layouts"
-import { SEO } from "components/media-and-icons"
-import { NextLinkButton } from "components/navigation"
-import { CONSTANT } from "constant"
-import { useI18n } from "contexts/i18n-context"
-import { PageProvider } from "contexts/page-context"
-import { TopLayout } from "layouts/top-layout"
 import { getStaticCommonProps } from "utils/next"
 import { Authentication } from "./authentication"
 import { Cards } from "./cards"
@@ -39,7 +39,6 @@ import { Mail } from "./mail"
 import { Music } from "./music"
 import { Playground } from "./playground"
 import { Tasks } from "./tasks"
-import { ArrowRight, Blocks } from "@yamada-ui/lucide"
 
 type Paths = GetStaticPathsResult["paths"]
