+++ docs/pages/examples/authentication/company.tsx
@@ -1,6 +1,6 @@
+import { Mountain } from "@yamada-ui/lucide"
 import { Heading, HStack, Spacer, Text, VStack } from "@yamada-ui/react"
 import type { StackProps } from "@yamada-ui/react"
-import { Mountain } from "@yamada-ui/lucide"
 import { memo } from "react"
 import type { FC } from "react"
