+++ docs/pages/examples/dashboard/index.tsx
@@ -1,4 +1,5 @@
 import { RangeDatePicker } from "@yamada-ui/calendar"
+import { ChevronDown } from "@yamada-ui/lucide"
 import {
   Button,
   Grid,
@@ -14,7 +15,6 @@ import {
   VStack,
 } from "@yamada-ui/react"
 import type { SegmentedControlItem, StackProps } from "@yamada-ui/react"
-import { ChevronDown } from "@yamada-ui/lucide"
 import { memo } from "react"
 import type { FC } from "react"
 import { Cards } from "./cards"
