+++ docs/pages/examples/mail/sidebar.tsx
@@ -1,15 +1,3 @@
-import type { Component, IconProps } from "@yamada-ui/react"
-import {
-  Button,
-  Divider,
-  HStack,
-  Heading,
-  IconButton,
-  Spacer,
-  Text,
-  Tooltip,
-  VStack,
-} from "@yamada-ui/react"
 import {
   Archive,
   CircleAlert,
@@ -22,6 +10,18 @@ import {
   Trash,
   Users,
 } from "@yamada-ui/lucide"
+import type { Component, IconProps } from "@yamada-ui/react"
+import {
+  Button,
+  Divider,
+  HStack,
+  Heading,
+  IconButton,
+  Spacer,
+  Text,
+  Tooltip,
+  VStack,
+} from "@yamada-ui/react"
 import { type FC, memo } from "react"
 import { Header } from "./header"
