+++ docs/pages/examples/music/sidebar.tsx
@@ -1,4 +1,13 @@
 import {
+  CirclePlay,
+  Globe,
+  ListMusic,
+  MicVocal,
+  Music2,
+  Radio,
+  User,
+} from "@yamada-ui/lucide"
+import {
   Box,
   Button,
   ButtonGroup,
@@ -17,15 +26,6 @@ import type {
   HeadingProps,
   StackProps,
 } from "@yamada-ui/react"
-import {
-  CirclePlay,
-  Globe,
-  ListMusic,
-  MicVocal,
-  Music2,
-  Radio,
-  User,
-} from "@yamada-ui/lucide"
 import { memo, useMemo } from "react"
 import type { ElementType, FC, ReactElement } from "react"
