+++ docs/pages/examples/tasks/filter.tsx
@@ -1,3 +1,4 @@
+import { CirclePlus, SlidersHorizontal, X } from "@yamada-ui/lucide"
 import {
   Button,
   Divider,
@@ -23,7 +24,6 @@ import type {
   InputProps,
   MenuProps,
 } from "@yamada-ui/react"
-import { CirclePlus, SlidersHorizontal, X } from "@yamada-ui/lucide"
 import { memo, useCallback, useMemo, useRef, useState } from "react"
 import type { FC, MutableRefObject, ReactNode } from "react"
 import { PRIORITY, STATUS, VIEW } from "./data"
