+++ docs/pages/icons/icon-drawer.tsx
@@ -1,3 +1,5 @@
+import type { IconNames } from "@yamada-ui/lucide"
+import { icons } from "@yamada-ui/lucide"
 import {
   assignRef,
   Center,
@@ -11,11 +13,9 @@ import {
   VStack,
 } from "@yamada-ui/react"
 import type { IconProps } from "@yamada-ui/react"
+import { CodeBlock } from "components/mdx/code-block"
 import { memo, useState } from "react"
 import type { FC, MutableRefObject } from "react"
-import type { IconNames } from "@yamada-ui/lucide"
-import { icons } from "@yamada-ui/lucide"
-import { CodeBlock } from "components/mdx/code-block"
 
 export type IconDrawerProps = {
   openRef: MutableRefObject<(name: IconNames) => void>
