+++ docs/pages/icons/list.tsx
@@ -1,3 +1,6 @@
+import type { IconNames } from "@yamada-ui/lucide"
+import { icons, Search as SearchIcon } from "@yamada-ui/lucide"
+import type { InputGroupProps, InputProps, StackProps } from "@yamada-ui/react"
 import {
   AspectRatio,
   assignRef,
@@ -12,16 +15,13 @@ import {
   Tooltip,
   VStack,
 } from "@yamada-ui/react"
-import type { InputGroupProps, InputProps, StackProps } from "@yamada-ui/react"
-import { useI18n } from "contexts/i18n-context"
-import { memo, useCallback, useMemo, useRef, useState } from "react"
-import type { FC, MutableRefObject } from "react"
-import type { IconNames } from "@yamada-ui/lucide"
-import { icons, Search as SearchIcon } from "@yamada-ui/lucide"
-import { TAGS } from "./tags"
+import { useI18n } from "contexts"
 import { matchSorter } from "match-sorter"
-import { NotFound } from "./not-found"
+import type { FC, MutableRefObject } from "react"
+import { memo, useCallback, useMemo, useRef, useState } from "react"
 import { IconDrawer } from "./icon-drawer"
+import { NotFound } from "./not-found"
+import { TAGS } from "./tags"
 
 const resolvedIcons = Object.entries(icons).filter(
   ([name]) => !name.endsWith("Icon"),
