+++ docs/pages/icons/tags.ts
@@ -1,8 +1,15 @@
 export const TAGS: Record<string, string[]> = {
-  AArrowUp: ["letter", "font size", "text", "formatting", "larger", "bigger"],
   ALargeSmall: ["letter", "font size", "text", "formatting"],
-  Accessibility: ["disability", "disabled", "dda", "wheelchair"],
+  AArrowUp: ["letter", "font size", "text", "formatting", "larger", "bigger"],
   AArrowDown: ["letter", "font size", "text", "formatting", "smaller"],
+  Accessibility: ["disability", "disabled", "dda", "wheelchair"],
+  AirVent: [
+    "air conditioner",
+    "ac",
+    "central air",
+    "cooling",
+    "climate-control",
+  ],
   Activity: [
     "pulse",
     "action",
@@ -32,17 +39,10 @@ export const TAGS: Record<string, string[]> = {
     "synthesiser",
     "music",
   ],
-  AirVent: [
-    "air conditioner",
-    "ac",
-    "central air",
-    "cooling",
-    "climate-control",
-  ],
   Airplay: ["stream", "cast", "mirroring", "screen", "monitor", "macos", "osx"],
   AlarmClockCheck: ["done", "todo", "tick", "complete", "task"],
-  AlarmClockOff: ["morning", "turn-off"],
   AlarmClockMinus: ["remove"],
+  AlarmClockOff: ["morning", "turn-off"],
   AlarmClockPlus: ["add"],
   AlarmClock: ["morning"],
   AlarmSmoke: [
@@ -55,8 +55,8 @@ export const TAGS: Record<string, string[]> = {
     "equipment",
     "amenities",
   ],
-  Album: ["photo", "book"],
   AlignCenterHorizontal: ["items", "flex", "justify"],
+  Album: ["photo", "book"],
   AlignCenterVertical: ["items", "flex", "justify"],
   AlignCenter: ["text", "alignment", "center"],
   AlignEndHorizontal: ["items", "bottom", "flex", "justify"],
@@ -73,7 +73,6 @@ export const TAGS: Record<string, string[]> = {
   AlignHorizontalDistributeStart: ["left", "items", "flex", "justify"],
   AlignHorizontalJustifyCenter: ["center", "items", "flex", "justify"],
   AlignHorizontalJustifyEnd: ["right", "items", "flex", "justify"],
-  AlignHorizontalJustifyStart: ["left", "items", "flex", "justify"],
   AlignHorizontalSpaceAround: [
     "center",
     "items",
@@ -82,13 +81,12 @@ export const TAGS: Record<string, string[]> = {
     "distribute",
     "between",
   ],
+  AlignHorizontalJustifyStart: ["left", "items", "flex", "justify"],
   AlignHorizontalSpaceBetween: ["around", "items", "bottom", "flex", "justify"],
   AlignJustify: ["text", "alignment", "justified", "menu", "list"],
   AlignLeft: ["text", "alignment", "left", "list"],
   AlignRight: ["text", "alignment", "right"],
-  AlignStartHorizontal: ["top", "items", "flex", "justify"],
   AlignStartVertical: ["left", "items", "flex", "justify"],
-  AlignVerticalDistributeEnd: ["bottom", "items", "flex", "justify"],
   AlignVerticalDistributeCenter: [
     "items",
     "flex",
@@ -97,6 +95,8 @@ export const TAGS: Record<string, string[]> = {
     "evenly",
     "around",
   ],
+  AlignStartHorizontal: ["top", "items", "flex", "justify"],
+  AlignVerticalDistributeEnd: ["bottom", "items", "flex", "justify"],
   AlignVerticalDistributeStart: ["top", "items", "flex", "justify"],
   AlignVerticalJustifyCenter: [
     "center",
@@ -161,8 +161,8 @@ export const TAGS: Record<string, string[]> = {
   ],
   Ampersands: ["and", "operator", "then", "code", "&&"],
   Anchor: ["ship"],
-  Angry: ["emoji", "anger", "face", "emotion"],
   Annoyed: ["emoji", "nuisance", "face", "emotion"],
+  Angry: ["emoji", "anger", "face", "emotion"],
   Antenna: [
     "signal",
     "connection",
@@ -204,10 +204,10 @@ export const TAGS: Record<string, string[]> = {
     "osx",
     "executable",
   ],
-  AppWindow: ["application", "menu bar", "pane", "executable"],
   Apple: ["fruit", "food"],
-  ArchiveX: ["index", "backup", "box", "storage", "records", "junk"],
+  AppWindow: ["application", "menu bar", "pane", "executable"],
   ArchiveRestore: ["unarchive", "index", "backup", "box", "storage", "records"],
+  ArchiveX: ["index", "backup", "box", "storage", "records", "junk"],
   Archive: ["index", "backup", "box", "storage", "records"],
   AreaChart: ["statistics", "analytics", "diagram", "graph", "area"],
   Armchair: ["sofa", "furniture", "leisure", "lounge", "loveseat", "couch"],
@@ -220,32 +220,10 @@ export const TAGS: Record<string, string[]> = {
     "download",
   ],
   ArrowBigDown: ["backwards", "reverse", "direction", "south"],
-  ArrowBigLeft: ["previous", "back", "direction", "west", "indicate turn"],
   ArrowBigLeftDash: ["previous", "back", "direction", "west", "turn", "corner"],
+  ArrowBigLeft: ["previous", "back", "direction", "west", "indicate turn"],
   ArrowBigRightDash: ["next", "forward", "direction", "east", "turn", "corner"],
   ArrowBigRight: ["next", "forward", "direction", "east", "indicate turn"],
-  ArrowBigUp: [
-    "shift",
-    "keyboard",
-    "button",
-    "mac",
-    "capitalize",
-    "capitalise",
-    "forward",
-    "direction",
-    "north",
-  ],
-  ArrowDown01: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-    "numerical",
-  ],
   ArrowBigUpDash: [
     "caps lock",
     "capitals",
@@ -259,28 +237,20 @@ export const TAGS: Record<string, string[]> = {
     "speed",
     "boost",
   ],
-  ArrowDown10: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-    "numerical",
+  ArrowBigUp: [
+    "shift",
+    "keyboard",
+    "button",
+    "mac",
+    "capitalize",
+    "capitalise",
+    "forward",
+    "direction",
+    "north",
   ],
-  ArrowDownAZ: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-    "alphabetical",
-  ],
+  ArrowDown01: ["filter", "sort", "ascending", "numerical"],
+  ArrowDown10: ["filter", "sort", "descending", "numerical"],
+  ArrowDownAZ: ["filter", "sort", "ascending", "alphabetical"],
   ArrowDownFromLine: [
     "backwards",
     "reverse",
@@ -292,16 +262,7 @@ export const TAGS: Record<string, string[]> = {
     "vertical",
   ],
   ArrowDownLeft: ["direction", "south-west", "diagonal"],
-  ArrowDownNarrowWide: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-  ],
+  ArrowDownNarrowWide: ["filter", "sort", "ascending"],
   ArrowDownRight: ["direction", "south-east", "diagonal"],
   ArrowDownToDot: [
     "direction",
@@ -310,6 +271,19 @@ export const TAGS: Record<string, string[]> = {
     "location",
     "step",
     "into",
+  ],
+  ArrowDownToLine: [
+    "behind",
+    "direction",
+    "south",
+    "download",
+    "save",
+    "git",
+    "version control",
+    "pull",
+    "collapse",
+    "fold",
+    "vertical",
   ],
   ArrowDownUp: [
     "bidirectional",
@@ -326,41 +300,8 @@ export const TAGS: Record<string, string[]> = {
     "reorder",
     "move",
   ],
-  ArrowDownToLine: [
-    "behind",
-    "direction",
-    "south",
-    "download",
-    "save",
-    "git",
-    "version control",
-    "pull",
-    "collapse",
-    "fold",
-    "vertical",
-  ],
-  ArrowDownWideNarrow: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-  ],
-  ArrowDownZA: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-    "alphabetical",
-    "reverse",
-  ],
+  ArrowDownWideNarrow: ["filter", "sort", "descending"],
+  ArrowDownZA: ["filter", "sort", "descending", "alphabetical", "reverse"],
   ArrowDown: ["backwards", "reverse", "direction", "south"],
   ArrowLeftFromLine: [
     "previous",
@@ -418,7 +359,6 @@ export const TAGS: Record<string, string[]> = {
     "<-",
     "->",
   ],
-  ArrowRight: ["forward", "next", "direction", "east", "->"],
   ArrowRightToLine: [
     "next",
     "forward",
@@ -433,28 +373,10 @@ export const TAGS: Record<string, string[]> = {
     "horizontal",
     "->|",
   ],
-  ArrowUp01: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-    "numerical",
-  ],
-  ArrowUp10: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-    "numerical",
-  ],
+  ArrowRight: ["forward", "next", "direction", "east", "->"],
+  ArrowUp01: ["filter", "sort", "ascending", "numerical"],
+  ArrowUp10: ["filter", "sort", "descending", "numerical"],
+  ArrowUpAZ: ["filter", "sort", "ascending", "alphabetical"],
   ArrowUpDown: [
     "bidirectional",
     "two-way",
@@ -467,17 +389,6 @@ export const TAGS: Record<string, string[]> = {
     "sort",
     "reorder",
     "move",
-  ],
-  ArrowUpAZ: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-    "alphabetical",
   ],
   ArrowUpFromDot: ["direction", "north", "step", "out"],
   ArrowUpFromLine: [
@@ -493,17 +404,8 @@ export const TAGS: Record<string, string[]> = {
     "vertical",
   ],
   ArrowUpLeft: ["direction", "north-west", "diagonal"],
+  ArrowUpNarrowWide: ["filter", "sort", "ascending"],
   ArrowUpRight: ["direction", "north-east", "diagonal"],
-  ArrowUpNarrowWide: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-  ],
   ArrowUpToLine: [
     "forward",
     "direction",
@@ -513,29 +415,9 @@ export const TAGS: Record<string, string[]> = {
     "fold",
     "vertical",
   ],
-  ArrowUpWideNarrow: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-  ],
-  ArrowUpZA: [
-    "filter",
-    "sort",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-    "alphabetical",
-    "reverse",
-  ],
+  ArrowUpWideNarrow: ["filter", "sort", "descending"],
   ArrowUp: ["forward", "direction", "north"],
+  ArrowUpZA: ["filter", "sort", "descending", "alphabetical", "reverse"],
   ArrowsUpFromLine: [
     "direction",
     "orientation",
@@ -547,7 +429,6 @@ export const TAGS: Record<string, string[]> = {
     "postage",
     "shipping",
   ],
-  AtSign: ["mention", "at", "email", "message", "@"],
   Asterisk: [
     "reference",
     "times",
@@ -559,6 +440,7 @@ export const TAGS: Record<string, string[]> = {
     "wildcard",
     "*",
   ],
+  AtSign: ["mention", "at", "email", "message", "@"],
   Atom: [
     "atomic",
     "nuclear",
@@ -637,6 +519,15 @@ export const TAGS: Record<string, string[]> = {
   ],
   Baby: ["child", "childproof", "children"],
   Backpack: ["bag", "hiking", "travel", "camping", "school", "childhood"],
+  BadgeAlert: [
+    "check",
+    "verified",
+    "unverified",
+    "security",
+    "safety",
+    "issue",
+  ],
+  BadgeCheck: ["verified", "check"],
   BadgeCent: [
     "discount",
     "offer",
@@ -655,15 +546,6 @@ export const TAGS: Record<string, string[]> = {
     "usd",
     "$",
     "¢",
-  ],
-  BadgeCheck: ["verified", "check"],
-  BadgeAlert: [
-    "check",
-    "verified",
-    "unverified",
-    "security",
-    "safety",
-    "issue",
   ],
   BadgeDollarSign: [
     "discount",
@@ -714,7 +596,6 @@ export const TAGS: Record<string, string[]> = {
     "₹",
   ],
   BadgeInfo: ["verified", "unverified", "help"],
-  BadgeMinus: ["verified", "unverified", "delete", "remove", "erase"],
   BadgeJapaneseYen: [
     "discount",
     "offer",
@@ -731,6 +612,7 @@ export const TAGS: Record<string, string[]> = {
     "jpy",
     "¥",
   ],
+  BadgeMinus: ["verified", "unverified", "delete", "remove", "erase"],
   BadgePercent: [
     "verified",
     "unverified",
@@ -792,8 +674,8 @@ export const TAGS: Record<string, string[]> = {
     "₣",
   ],
   BadgeX: ["verified", "unverified", "lost", "delete", "remove"],
-  BaggageClaim: ["baggage", "luggage", "travel", "cart", "trolley", "suitcase"],
   Badge: ["check", "verified", "unverified"],
+  BaggageClaim: ["baggage", "luggage", "travel", "cart", "trolley", "suitcase"],
   Ban: [
     "cancel",
     "no",
@@ -874,7 +756,6 @@ export const TAGS: Record<string, string[]> = {
   ],
   BeerOff: ["alcohol", "bar", "beverage", "brewery", "drink"],
   Beer: ["alcohol", "bar", "beverage", "brewery", "drink"],
-  BellDot: ["alarm", "notification", "sound", "reminder", "unread"],
   BellElectric: [
     "fire alarm",
     "flames",
@@ -896,6 +777,7 @@ export const TAGS: Record<string, string[]> = {
     "ring",
     "reception",
   ],
+  BellDot: ["alarm", "notification", "sound", "reminder", "unread"],
   BellMinus: [
     "alarm",
     "notification",
@@ -907,8 +789,8 @@ export const TAGS: Record<string, string[]> = {
   ],
   BellOff: ["alarm", "notification", "silent", "reminder"],
   BellPlus: ["notification", "silent", "reminder", "add", "create", "new"],
-  Bell: ["alarm", "notification", "sound", "reminder"],
   BellRing: ["alarm", "notification", "sound", "reminder"],
+  Bell: ["alarm", "notification", "sound", "reminder"],
   BetweenHorizontalEnd: [
     "insert",
     "add",
@@ -955,26 +837,6 @@ export const TAGS: Record<string, string[]> = {
     "rectangles",
     "chevron",
   ],
-  BetweenVerticalStart: [
-    "insert",
-    "add",
-    "bottom",
-    "slot",
-    "squeeze",
-    "space",
-    "vertical",
-    "grid",
-    "table",
-    "columns",
-    "cells",
-    "data",
-    "enter",
-    "entry",
-    "entries",
-    "blocks",
-    "rectangles",
-    "chevron",
-  ],
   BetweenVerticalEnd: [
     "insert",
     "add",
@@ -995,6 +857,27 @@ export const TAGS: Record<string, string[]> = {
     "rectangles",
     "chevron",
   ],
+  BetweenVerticalStart: [
+    "insert",
+    "add",
+    "bottom",
+    "slot",
+    "squeeze",
+    "space",
+    "vertical",
+    "grid",
+    "table",
+    "columns",
+    "cells",
+    "data",
+    "enter",
+    "entry",
+    "entries",
+    "blocks",
+    "rectangles",
+    "chevron",
+  ],
+  Bike: ["bicycle", "transport", "trip"],
   BicepsFlexed: [
     "arm",
     "muscle",
@@ -1006,9 +889,7 @@ export const TAGS: Record<string, string[]> = {
     "forelimb",
     "curled",
   ],
-  Bike: ["bicycle", "transport", "trip"],
   Binary: ["code", "digits", "computer", "zero", "one", "boolean"],
-  Bird: ["peace", "freedom", "wing", "avian", "tweet"],
   Biohazard: [
     "fallout",
     "waste",
@@ -1018,25 +899,17 @@ export const TAGS: Record<string, string[]> = {
     "element",
   ],
   Bitcoin: ["currency", "money", "payment"],
-  Blend: [
-    "mode",
-    "overlay",
-    "multiply",
-    "screen",
-    "opacity",
-    "transparency",
-    "alpha",
-    "filters",
-    "lenses",
-    "mixed",
+  Bird: ["peace", "freedom", "wing", "avian", "tweet"],
+  Blinds: [
     "shades",
-    "tints",
-    "hues",
-    "saturation",
-    "brightness",
-    "overlap",
-    "colors",
-    "colours",
+    "screen",
+    "curtain",
+    "shutter",
+    "roller blind",
+    "window",
+    "lighting",
+    "household",
+    "home",
   ],
   Blocks: [
     "addon",
@@ -1051,22 +924,22 @@ export const TAGS: Record<string, string[]> = {
     "children",
     "learning",
   ],
-  Blinds: [
-    "shades",
-    "screen",
-    "curtain",
-    "shutter",
-    "roller blind",
-    "window",
-    "lighting",
-    "household",
-    "home",
-  ],
   BluetoothConnected: ["paired"],
-  BluetoothSearching: ["pairing"],
   BluetoothOff: ["lost"],
-  Bluetooth: ["wireless"],
+  BluetoothSearching: ["pairing"],
   Bold: ["text", "strong", "format"],
+  Bluetooth: ["wireless"],
+  Bomb: [
+    "fatal",
+    "error",
+    "crash",
+    "blockbuster",
+    "mine",
+    "explosion",
+    "explode",
+    "explosive",
+  ],
+  Bone: ["health", "skeleton", "skull", "death", "pets", "dog"],
   Bolt: [
     "nut",
     "screw",
@@ -1081,17 +954,6 @@ export const TAGS: Record<string, string[]> = {
     "contruction",
     "parts",
   ],
-  Bomb: [
-    "fatal",
-    "error",
-    "crash",
-    "blockbuster",
-    "mine",
-    "explosion",
-    "explode",
-    "explosive",
-  ],
-  Bone: ["health", "skeleton", "skull", "death", "pets", "dog"],
   BookA: [
     "dictionary",
     "define",
@@ -1240,6 +1102,26 @@ export const TAGS: Record<string, string[]> = {
     "unpublished",
     "untold",
   ],
+  Blend: [
+    "mode",
+    "overlay",
+    "multiply",
+    "screen",
+    "opacity",
+    "transparency",
+    "alpha",
+    "filters",
+    "lenses",
+    "mixed",
+    "shades",
+    "tints",
+    "hues",
+    "saturation",
+    "brightness",
+    "overlap",
+    "colors",
+    "colours",
+  ],
   BookDown: ["code", "coding", "version control", "git", "repository", "pull"],
   BookHeadphones: [
     "audiobook",
@@ -1256,6 +1138,20 @@ export const TAGS: Record<string, string[]> = {
     "study",
     "learning",
     "research",
+  ],
+  BookImage: [
+    "images",
+    "pictures",
+    "photos",
+    "album",
+    "collection",
+    "event",
+    "magazine",
+    "catalog",
+    "catalogue",
+    "brochure",
+    "browse",
+    "gallery",
   ],
   BookHeart: [
     "diary",
@@ -1291,20 +1187,6 @@ export const TAGS: Record<string, string[]> = {
     "favorite",
     "favourite",
     "high school",
-  ],
-  BookImage: [
-    "images",
-    "pictures",
-    "photos",
-    "album",
-    "collection",
-    "event",
-    "magazine",
-    "catalog",
-    "catalogue",
-    "brochure",
-    "browse",
-    "gallery",
   ],
   BookKey: [
     "code",
@@ -1594,6 +1476,7 @@ export const TAGS: Record<string, string[]> = {
     "push",
     "force",
   ],
+  BookUp: ["code", "coding", "version control", "git", "repository", "push"],
   BookUser: [
     "person",
     "people",
@@ -1608,7 +1491,6 @@ export const TAGS: Record<string, string[]> = {
     "listing",
     "networking",
   ],
-  BookUp: ["code", "coding", "version control", "git", "repository", "push"],
   BookX: [
     "code",
     "coding",
@@ -1635,16 +1517,6 @@ export const TAGS: Record<string, string[]> = {
     "uneducated",
     "re-education",
     "unlearn",
-  ],
-  BookmarkCheck: [
-    "read",
-    "finished",
-    "complete",
-    "clip",
-    "marker",
-    "tag",
-    "task",
-    "todo",
   ],
   Book: [
     "reading",
@@ -1678,6 +1550,16 @@ export const TAGS: Record<string, string[]> = {
     "documentation",
   ],
   BookmarkMinus: ["delete", "remove"],
+  BookmarkCheck: [
+    "read",
+    "finished",
+    "complete",
+    "clip",
+    "marker",
+    "tag",
+    "task",
+    "todo",
+  ],
   BookmarkPlus: ["add"],
   BookmarkX: [
     "read",
@@ -1701,6 +1583,7 @@ export const TAGS: Record<string, string[]> = {
     "live",
     "frequency",
   ],
+  Bot: ["robot", "ai", "chat", "assistant"],
   BoxSelect: [
     "selection",
     "square",
@@ -1728,7 +1611,6 @@ export const TAGS: Record<string, string[]> = {
     "cluster",
     "geometry",
   ],
-  BotOff: ["robot", "ai", "chat", "assistant"],
   Braces: ["json", "code", "token", "curly brackets", "data", "{", "}"],
   Brackets: ["code", "token", "array", "list", "square", "[", "]"],
   BrainCircuit: [
@@ -1798,9 +1680,9 @@ export const TAGS: Record<string, string[]> = {
     "organization",
   ],
   Building: ["organisation", "organization"],
-  BusFront: ["coach", "vehicle", "trip", "road"],
-  Bus: ["bus", "vehicle", "transport", "trip"],
+  BotOff: ["robot", "ai", "chat", "assistant"],
   CableCar: ["ski lift", "winter holiday", "alpine", "resort", "mountains"],
+  Bus: ["bus", "vehicle", "transport", "trip"],
   Cable: [
     "cord",
     "wire",
@@ -1832,19 +1714,6 @@ export const TAGS: Record<string, string[]> = {
     "television",
     "optical",
   ],
-  Cake: [
-    "birthday",
-    "birthdate",
-    "celebration",
-    "party",
-    "surprise",
-    "gateaux",
-    "dessert",
-    "fondant",
-    "icing sugar",
-    "sweet",
-    "baking",
-  ],
   CakeSlice: [
     "birthday",
     "birthdate",
@@ -1860,49 +1729,18 @@ export const TAGS: Record<string, string[]> = {
     "sweet",
     "baking",
   ],
-  Calculator: ["count", "calculating machine"],
-  CalendarArrowDown: [
-    "date",
-    "month",
-    "year",
-    "event",
-    "sort",
-    "order",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-  ],
-  CalendarCheck: [
-    "date",
-    "day",
-    "month",
-    "year",
-    "event",
-    "confirm",
-    "subscribe",
-    "schedule",
-    "done",
-    "todo",
-    "tick",
-    "complete",
-    "task",
-  ],
-  CalendarArrowUp: [
-    "date",
-    "month",
-    "year",
-    "event",
-    "sort",
-    "order",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
+  Cake: [
+    "birthday",
+    "birthdate",
+    "celebration",
+    "party",
+    "surprise",
+    "gateaux",
+    "dessert",
+    "fondant",
+    "icing sugar",
+    "sweet",
+    "baking",
   ],
   CalendarCheck2: [
     "date",
@@ -1919,7 +1757,22 @@ export const TAGS: Record<string, string[]> = {
     "complete",
     "task",
   ],
-  CalendarClock: ["date", "day", "month", "year", "event", "clock", "hour"],
+  Calculator: ["count", "calculating machine"],
+  CalendarCheck: [
+    "date",
+    "day",
+    "month",
+    "year",
+    "event",
+    "confirm",
+    "subscribe",
+    "schedule",
+    "done",
+    "todo",
+    "tick",
+    "complete",
+    "task",
+  ],
   CalendarCog: [
     "date",
     "day",
@@ -1930,16 +1783,8 @@ export const TAGS: Record<string, string[]> = {
     "gear",
     "cog",
   ],
+  CalendarClock: ["date", "day", "month", "year", "event", "clock", "hour"],
   CalendarDays: ["date", "month", "year", "event"],
-  CalendarFold: [
-    "date",
-    "month",
-    "year",
-    "event",
-    "birthday",
-    "birthdate",
-    "ics",
-  ],
   CalendarHeart: [
     "date",
     "month",
@@ -1952,17 +1797,6 @@ export const TAGS: Record<string, string[]> = {
   ],
   CalendarMinus2: ["date", "day", "month", "year", "event", "delete", "remove"],
   CalendarOff: ["date", "day", "month", "year", "event", "delete", "remove"],
-  CalendarPlus: [
-    "date",
-    "day",
-    "month",
-    "year",
-    "event",
-    "add",
-    "subscribe",
-    "create",
-    "new",
-  ],
   CalendarMinus: ["date", "day", "month", "year", "event", "delete", "remove"],
   CalendarPlus2: [
     "date",
@@ -1975,12 +1809,32 @@ export const TAGS: Record<string, string[]> = {
     "create",
     "new",
   ],
+  CalendarPlus: [
+    "date",
+    "day",
+    "month",
+    "year",
+    "event",
+    "add",
+    "subscribe",
+    "create",
+    "new",
+  ],
   CalendarRange: ["date", "day", "month", "year", "event", "range", "period"],
   CalendarSearch: ["date", "day", "month", "year", "events", "search"],
+  CalendarFold: [
+    "date",
+    "month",
+    "year",
+    "event",
+    "birthday",
+    "birthdate",
+    "ics",
+  ],
+  CalendarX2: ["date", "day", "month", "year", "event", "remove"],
   CalendarX: ["date", "day", "month", "year", "event", "remove", "busy"],
   Calendar: ["date", "month", "year", "event", "birthday", "birthdate"],
   CameraOff: ["photo", "webcam", "video"],
-  CalendarX2: ["date", "day", "month", "year", "event", "remove"],
   Camera: ["photo", "webcam", "video"],
   CandlestickChart: [
     "trading",
@@ -2004,8 +1858,8 @@ export const TAGS: Record<string, string[]> = {
   ],
   CandyCane: ["sugar", "food", "sweet", "christmas", "xmas"],
   CandyOff: ["sugar free", "food", "sweet", "allergy", "intolerance", "diet"],
-  Candy: ["sugar", "food", "sweet"],
   Cannabis: ["cannabis", "weed", "leaf"],
+  Candy: ["sugar", "food", "sweet"],
   CaptionsOff: [
     "closed captions",
     "subtitles",
@@ -2015,7 +1869,6 @@ export const TAGS: Record<string, string[]> = {
     "dialogue",
     "accessibility",
   ],
-  CarFront: ["vehicle", "drive", "trip", "journey"],
   Captions: [
     "closed captions",
     "subtitles",
@@ -2025,7 +1878,9 @@ export const TAGS: Record<string, string[]> = {
     "dialogue",
     "accessibility",
   ],
+  CarFront: ["vehicle", "drive", "trip", "journey"],
   CarTaxiFront: ["cab", "vehicle", "drive", "trip", "journey"],
+  Car: ["vehicle", "drive", "trip", "journey"],
   Caravan: [
     "trailer",
     "tow",
@@ -2039,13 +1894,21 @@ export const TAGS: Record<string, string[]> = {
   ],
   Carrot: ["vegetable", "food", "eat"],
   CaseLower: ["text", "letters", "characters", "font", "typography"],
-  CaseSensitive: ["text", "letters", "characters", "font", "typography"],
   CaseUpper: ["text", "letters", "characters", "font", "typography"],
+  CaseSensitive: ["text", "letters", "characters", "font", "typography"],
   Cast: ["chromecast", "airplay", "screen"],
   CassetteTape: ["audio", "music", "recording", "play"],
-  Castle: ["fortress", "stronghold", "palace", "chateau", "building"],
   Cat: ["animal", "pet", "kitten", "feline"],
-  Car: ["vehicle", "drive", "trip", "journey"],
+  Castle: ["fortress", "stronghold", "palace", "chateau", "building"],
+  Cctv: [
+    "camera",
+    "surveillance",
+    "recording",
+    "film",
+    "videotape",
+    "crime",
+    "watching",
+  ],
   CheckCheck: [
     "done",
     "received",
@@ -2056,18 +1919,11 @@ export const TAGS: Record<string, string[]> = {
     "task",
   ],
   Check: ["done", "todo", "tick", "complete", "task"],
-  Cctv: [
-    "camera",
-    "surveillance",
-    "recording",
-    "film",
-    "videotape",
-    "crime",
-    "watching",
-  ],
-  ChefHat: ["cooking", "food", "kitchen", "restaurant"],
   Cherry: ["fruit", "food"],
+  ChefHat: ["cooking", "food", "kitchen", "restaurant"],
+  ChevronDown: ["backwards", "reverse", "slow", "dropdown"],
   ChevronFirst: ["previous", "music"],
+  ChevronLeft: ["back", "previous", "less than", "fewer", "menu", "<"],
   ChevronLast: ["skip", "next", "music"],
   ChevronRight: [
     "forward",
@@ -2083,9 +1939,6 @@ export const TAGS: Record<string, string[]> = {
     "shell",
     ">",
   ],
-  ChevronLeft: ["back", "previous", "less than", "fewer", "menu", "<"],
-  ChevronDown: ["backwards", "reverse", "slow", "dropdown"],
-  ChevronsDownUp: ["collapse", "fold", "vertical"],
   ChevronUp: [
     "caret",
     "keyboard",
@@ -2100,14 +1953,15 @@ export const TAGS: Record<string, string[]> = {
     "^",
     "dropdown",
   ],
-  ChevronsDown: ["backwards", "reverse", "slower"],
+  ChevronsDownUp: ["collapse", "fold", "vertical"],
   ChevronsLeftRight: ["expand", "horizontal", "unfold"],
+  ChevronsDown: ["backwards", "reverse", "slower"],
   ChevronsLeft: ["turn", "corner"],
   ChevronsRightLeft: ["collapse", "fold", "horizontal"],
   ChevronsRight: ["turn", "corner"],
   ChevronsUpDown: ["expand", "unfold", "vertical"],
-  Chrome: ["browser", "logo"],
   ChevronsUp: ["forward", "ahead", "faster", "speed", "boost"],
+  Chrome: ["browser", "logo"],
   Church: ["temple", "building"],
   CigaretteOff: ["smoking", "no-smoking"],
   Cigarette: ["smoking"],
@@ -2131,6 +1985,8 @@ export const TAGS: Record<string, string[]> = {
     "<-",
   ],
   CircleArrowOutDownRight: ["outwards", "direction", "south-east", "diagonal"],
+  CircleArrowOutDownLeft: ["outwards", "direction", "south-west", "diagonal"],
+  CircleArrowOutUpRight: ["outwards", "direction", "north-east", "diagonal"],
   CircleArrowOutUpLeft: [
     "outwards",
     "direction",
@@ -2140,8 +1996,6 @@ export const TAGS: Record<string, string[]> = {
     "button",
     "escape",
   ],
-  CircleArrowOutDownLeft: ["outwards", "direction", "south-west", "diagonal"],
-  CircleArrowOutUpRight: ["outwards", "direction", "north-east", "diagonal"],
   CircleArrowRight: [
     "next",
     "forward",
@@ -2153,13 +2007,12 @@ export const TAGS: Record<string, string[]> = {
     "->",
   ],
   CircleArrowUp: ["forward", "direction", "north", "sign", "button"],
-  CircleCheckBig: ["done", "todo", "tick", "complete", "task"],
   CircleCheck: ["done", "todo", "tick", "complete", "task"],
+  CircleCheckBig: ["done", "todo", "tick", "complete", "task"],
   CircleChevronDown: ["back", "menu"],
   CircleChevronLeft: ["back", "previous", "less than", "fewer", "menu", "<"],
   CircleChevronRight: ["back", "more than", "greater", "menu", ">"],
   CircleChevronUp: ["caret", "ahead", "menu", "^"],
-  CircleDivide: ["calculate", "maths", "÷", "/"],
   CircleDashed: [
     "pending",
     "dot",
@@ -2170,6 +2023,7 @@ export const TAGS: Record<string, string[]> = {
     "coding",
     "version control",
   ],
+  CircleDivide: ["calculate", "maths", "÷", "/"],
   CircleDollarSign: [
     "monetization",
     "marketing",
@@ -2217,7 +2071,6 @@ export const TAGS: Record<string, string[]> = {
     "multiple choice",
     "choose",
   ],
-  CircleEqual: ["calculate", "shape", "="],
   CircleFadingPlus: [
     "stories",
     "social media",
@@ -2228,6 +2081,8 @@ export const TAGS: Record<string, string[]> = {
     "sharing",
     "content",
   ],
+  CircleEqual: ["calculate", "shape", "="],
+  CircleHelp: ["question mark"],
   CircleGauge: [
     "dashboard",
     "dial",
@@ -2251,7 +2106,7 @@ export const TAGS: Record<string, string[]> = {
     "downgrade",
     "-",
   ],
-  CircleHelp: ["question mark"],
+  CircleParking: ["parking lot", "car park"],
   CircleOff: [
     "diameter",
     "zero",
@@ -2272,7 +2127,7 @@ export const TAGS: Record<string, string[]> = {
     "failure",
   ],
   CircleParkingOff: ["parking lot", "car park", "no parking"],
-  CircleParking: ["parking lot", "car park"],
+  CirclePause: ["music", "audio", "stop"],
   CirclePercent: [
     "verified",
     "unverified",
@@ -2283,7 +2138,6 @@ export const TAGS: Record<string, string[]> = {
     "sticker",
     "price tag",
   ],
-  CirclePause: ["music", "audio", "stop"],
   CirclePlay: ["music", "start", "run"],
   CirclePower: [
     "on",
@@ -2360,9 +2214,7 @@ export const TAGS: Record<string, string[]> = {
     "or",
     "/",
   ],
-  CircleUser: ["person", "account", "contact"],
   CircleStop: ["media", "music"],
-  CircleUserRound: ["person", "account", "contact"],
   CircleX: [
     "cancel",
     "close",
@@ -2379,9 +2231,20 @@ export const TAGS: Record<string, string[]> = {
     "multiply",
     "multiplication",
   ],
-  Citrus: ["lemon", "orange", "grapefruit", "fruit"],
+  CircleUserRound: ["person", "account", "contact"],
+  CircleUser: ["person", "account", "contact"],
   Circle: ["off", "zero", "record", "shape"],
+  Citrus: ["lemon", "orange", "grapefruit", "fruit"],
   CircuitBoard: ["computing", "electricity", "electronics"],
+  ClipboardCheck: [
+    "copied",
+    "pasted",
+    "done",
+    "todo",
+    "tick",
+    "complete",
+    "task",
+  ],
   Clapperboard: [
     "movie",
     "film",
@@ -2395,18 +2258,8 @@ export const TAGS: Record<string, string[]> = {
     "show",
     "entertainment",
   ],
-  ClipboardCheck: [
-    "copied",
-    "pasted",
-    "done",
-    "todo",
-    "tick",
-    "complete",
-    "task",
-  ],
-  ClipboardList: ["copy", "paste", "tasks"],
   ClipboardCopy: ["copy", "paste"],
-  ClipboardPaste: ["copy", "paste"],
+  ClipboardList: ["copy", "paste", "tasks"],
   ClipboardMinus: [
     "copy",
     "delete",
@@ -2417,7 +2270,9 @@ export const TAGS: Record<string, string[]> = {
     "report",
     "doctor",
   ],
+  ClipboardPaste: ["copy", "paste"],
   ClipboardPenLine: ["paste"],
+  ClipboardPen: ["paste", "signature"],
   ClipboardPlus: [
     "copy",
     "paste",
@@ -2429,74 +2284,47 @@ export const TAGS: Record<string, string[]> = {
     "report",
     "doctor",
   ],
-  ClipboardPen: ["paste", "signature"],
   ClipboardType: ["paste", "format", "text"],
   Clipboard: ["copy", "paste"],
-  Clock1: ["time", "watch", "alarm"],
   ClipboardX: ["copy", "paste", "discard", "remove"],
   Clock10: ["time", "watch", "alarm"],
+  Clock1: ["time", "watch", "alarm"],
   Clock11: ["time", "watch", "alarm"],
-  Clock12: ["time", "watch", "alarm", "noon", "midnight"],
+  BusFront: ["coach", "vehicle", "trip", "road"],
   Clock2: ["time", "watch", "alarm"],
   Clock3: ["time", "watch", "alarm"],
-  Clock5: ["time", "watch", "alarm"],
   Clock4: ["time", "watch", "alarm"],
-  Clock7: ["time", "watch", "alarm"],
+  Clock5: ["time", "watch", "alarm"],
   Clock6: ["time", "watch", "alarm"],
+  Clock7: ["time", "watch", "alarm"],
   Clock8: ["time", "watch", "alarm"],
-  ClockArrowDown: [
-    "time",
-    "watch",
-    "alarm",
-    "sort",
-    "order",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-  ],
   Clock9: ["time", "watch", "alarm"],
   CloudCog: ["computing", "ai", "cluster", "network"],
-  ClockArrowUp: [
-    "time",
-    "watch",
-    "alarm",
-    "sort",
-    "order",
-    "ascending",
-    "descending",
-    "increasing",
-    "decreasing",
-    "rising",
-    "falling",
-  ],
+  CloudDrizzle: ["weather", "shower"],
   Clock: ["time", "watch", "alarm"],
   CloudDownload: ["import"],
   CloudFog: ["weather", "mist"],
   CloudHail: ["weather", "rainfall"],
-  CloudDrizzle: ["weather", "shower"],
   CloudMoonRain: ["weather", "partly", "night", "rainfall"],
-  CloudOff: ["disconnect"],
   CloudLightning: ["weather", "bolt"],
+  CloudOff: ["disconnect"],
   CloudMoon: ["weather", "night"],
-  CloudRain: ["weather", "rainfall"],
   CloudRainWind: ["weather", "rainfall"],
-  CloudSunRain: ["weather", "partly", "rainfall"],
+  CloudRain: ["weather", "rainfall"],
   CloudSnow: ["weather", "blizzard"],
-  CloudUpload: ["file"],
+  CloudSunRain: ["weather", "partly", "rainfall"],
   CloudSun: ["weather", "partly"],
-  Cloudy: ["weather", "clouds"],
+  CloudUpload: ["file"],
   Cloud: ["weather"],
-  Club: ["shape", "suit", "playing", "cards"],
+  Cloudy: ["weather", "clouds"],
   Clover: ["leaf", "luck", "plant"],
+  Club: ["shape", "suit", "playing", "cards"],
   CodeXml: ["source", "programming", "html", "xml"],
   Code: ["source", "programming", "html", "xml"],
-  Codepen: ["logo"],
   Codesandbox: ["logo"],
-  Coffee: ["drink", "cup", "mug", "tea", "cafe", "hot", "beverage"],
+  Codepen: ["logo"],
   Cog: ["computing", "settings", "cog", "edit", "gear", "preferences"],
+  Coffee: ["drink", "cup", "mug", "tea", "cafe", "hot", "beverage"],
   Coins: ["money", "cash", "finance", "gamble"],
   Columns2: [
     "lines",
@@ -2560,17 +2388,6 @@ export const TAGS: Record<string, string[]> = {
     "enforcement",
     "grill",
   ],
-  Combine: [
-    "cubes",
-    "packages",
-    "parts",
-    "units",
-    "collection",
-    "cluster",
-    "combine",
-    "gather",
-    "merge",
-  ],
   Columns3: [
     "lines",
     "list",
@@ -2597,18 +2414,21 @@ export const TAGS: Record<string, string[]> = {
     "pages",
   ],
   Command: ["keyboard", "key", "mac", "cmd", "button"],
-  Compass: ["direction", "north", "east", "south", "west", "safari", "browser"],
-  Component: ["design", "element", "group", "module", "part", "symbol"],
-  Computer: ["pc", "chassis", "codespaces", "github"],
-  ContactRound: [
-    "user",
-    "person",
-    "family",
-    "friend",
-    "acquaintance",
-    "listing",
-    "networking",
+  Combine: [
+    "cubes",
+    "packages",
+    "parts",
+    "units",
+    "collection",
+    "cluster",
+    "combine",
+    "gather",
+    "merge",
   ],
+  Component: ["design", "element", "group", "module", "part", "symbol"],
+  Compass: ["direction", "north", "east", "south", "west", "safari", "browser"],
+  ConciergeBell: ["reception", "bell", "porter"],
+  Computer: ["pc", "chassis", "codespaces", "github"],
   Cone: [
     "conical",
     "triangle",
@@ -2620,8 +2440,25 @@ export const TAGS: Record<string, string[]> = {
     "spotlight",
     "searchlight",
   ],
-  ConciergeBell: ["reception", "bell", "porter"],
   Construction: ["roadwork", "maintenance", "blockade", "barricade"],
+  Contact: [
+    "user",
+    "person",
+    "family",
+    "friend",
+    "acquaintance",
+    "listing",
+    "networking",
+  ],
+  ContactRound: [
+    "user",
+    "person",
+    "family",
+    "friend",
+    "acquaintance",
+    "listing",
+    "networking",
+  ],
   Container: [
     "storage",
     "shipping",
@@ -2634,15 +2471,6 @@ export const TAGS: Record<string, string[]> = {
     "coding",
   ],
   Contrast: ["display", "accessibility"],
-  Contact: [
-    "user",
-    "person",
-    "family",
-    "friend",
-    "acquaintance",
-    "listing",
-    "networking",
-  ],
   Cookie: ["biscuit", "privacy", "legal", "food"],
   CookingPot: [
     "pod",
@@ -2658,6 +2486,7 @@ export const TAGS: Record<string, string[]> = {
     "meal",
     "eat",
   ],
+  CopyCheck: ["clone", "duplicate", "done", "multiple"],
   CopyMinus: [
     "clone",
     "duplicate",
@@ -2668,7 +2497,6 @@ export const TAGS: Record<string, string[]> = {
     "multiple",
     "-",
   ],
-  CopyCheck: ["clone", "duplicate", "done", "multiple"],
   CopyPlus: ["clone", "duplicate", "add", "multiple", "expand", "+"],
   CopySlash: [
     "clone",
@@ -2687,7 +2515,6 @@ export const TAGS: Record<string, string[]> = {
     "or",
     "/",
   ],
-  Copy: ["clone", "duplicate", "multiple"],
   CopyX: [
     "cancel",
     "close",
@@ -2699,18 +2526,16 @@ export const TAGS: Record<string, string[]> = {
     "multiplication",
     "times",
   ],
-  Copyleft: ["licence"],
+  Copy: ["clone", "duplicate", "multiple"],
   Copyright: ["licence", "license"],
-  Bot: ["robot", "ai", "chat", "assistant"],
-  CornerDownLeft: ["arrow", "return"],
-  CornerDownRight: ["arrow", "indent", "tab"],
+  Copyleft: ["licence"],
   CornerLeftDown: ["arrow"],
-  CornerRightDown: ["arrow"],
+  CornerDownRight: ["arrow", "indent", "tab"],
+  CornerDownLeft: ["arrow", "return"],
   CornerLeftUp: ["arrow"],
+  CornerRightDown: ["arrow"],
   CornerRightUp: ["arrow"],
   CornerUpLeft: ["arrow"],
-  CornerUpRight: ["arrow"],
-  CreativeCommons: ["licence", "license"],
   Cpu: [
     "processor",
     "cores",
@@ -2724,10 +2549,13 @@ export const TAGS: Record<string, string[]> = {
     "gigahertz",
     "ghz",
   ],
+  CornerUpRight: ["arrow"],
+  CreativeCommons: ["licence", "license"],
   CreditCard: ["bank", "purchase", "payment", "cc"],
   Croissant: ["bakery", "cooking", "food", "pastry"],
   Crop: ["photo", "image"],
   Crosshair: ["aim", "target"],
+  Cross: ["healthcare", "first aid"],
   Crown: [
     "diadem",
     "tiara",
@@ -2738,7 +2566,6 @@ export const TAGS: Record<string, string[]> = {
     "winner",
     "favourite",
   ],
-  Cross: ["healthcare", "first aid"],
   Cuboid: [
     "brick",
     "block",
@@ -2748,8 +2575,17 @@ export const TAGS: Record<string, string[]> = {
     "rectangular",
     "hexahedron",
   ],
-  CupSoda: ["beverage", "cup", "drink", "soda", "straw", "water"],
   Currency: ["finance", "money"],
+  CupSoda: ["beverage", "cup", "drink", "soda", "straw", "water"],
+  Cylinder: [
+    "shape",
+    "elliptical",
+    "geometry",
+    "container",
+    "storage",
+    "tin",
+    "pot",
+  ],
   DatabaseBackup: [
     "storage",
     "memory",
@@ -2761,14 +2597,14 @@ export const TAGS: Record<string, string[]> = {
     "arrow",
     "left",
   ],
-  Cylinder: [
-    "shape",
-    "elliptical",
-    "geometry",
-    "container",
+  Database: [
     "storage",
+    "memory",
+    "container",
     "tin",
     "pot",
+    "bytes",
+    "servers",
   ],
   DatabaseZap: [
     "cache busting",
@@ -2779,27 +2615,7 @@ export const TAGS: Record<string, string[]> = {
     "power",
     "crash",
   ],
-  Database: [
-    "storage",
-    "memory",
-    "container",
-    "tin",
-    "pot",
-    "bytes",
-    "servers",
-  ],
   Delete: ["backspace", "remove"],
-  Diameter: [
-    "shape",
-    "circle",
-    "geometry",
-    "trigonometry",
-    "width",
-    "height",
-    "size",
-    "calculate",
-    "measure",
-  ],
   Dessert: [
     "pudding",
     "christmas",
@@ -2843,6 +2659,17 @@ export const TAGS: Record<string, string[]> = {
     "minimum",
     "downgrade",
   ],
+  Diameter: [
+    "shape",
+    "circle",
+    "geometry",
+    "trigonometry",
+    "width",
+    "height",
+    "size",
+    "calculate",
+    "measure",
+  ],
   DiamondPercent: [
     "verified",
     "unverified",
@@ -2852,16 +2679,6 @@ export const TAGS: Record<string, string[]> = {
     "marketing",
     "sticker",
     "price tag",
-  ],
-  Diamond: [
-    "square",
-    "rectangle",
-    "oblique",
-    "rhombus",
-    "shape",
-    "suit",
-    "playing",
-    "cards",
   ],
   DiamondPlus: [
     "keyframe",
@@ -2884,10 +2701,22 @@ export const TAGS: Record<string, string[]> = {
     "+",
   ],
   Dice1: ["dice", "random", "tabletop", "1", "board", "game"],
+  Diamond: [
+    "square",
+    "rectangle",
+    "oblique",
+    "rhombus",
+    "shape",
+    "suit",
+    "playing",
+    "cards",
+  ],
   Dice2: ["dice", "random", "tabletop", "2", "board", "game"],
   Dice3: ["dice", "random", "tabletop", "3", "board", "game"],
+  Dice6: ["dice", "random", "tabletop", "6", "board", "game"],
   Dice4: ["dice", "random", "tabletop", "4", "board", "game"],
   Dice5: ["dice", "random", "tabletop", "5", "board", "game"],
+  Dices: ["dice", "random", "tabletop", "board", "game"],
   Diff: [
     "patch",
     "difference",
@@ -2897,8 +2726,6 @@ export const TAGS: Record<string, string[]> = {
     "plus-minus",
     "maths",
   ],
-  Dices: ["dice", "random", "tabletop", "board", "game"],
-  Dice6: ["dice", "random", "tabletop", "6", "board", "game"],
   Disc2: [
     "album",
     "music",
@@ -2965,7 +2792,16 @@ export const TAGS: Record<string, string[]> = {
     "rpm",
     "dj",
   ],
+  DnaOff: [
+    "gene",
+    "gmo free",
+    "helix",
+    "heredity",
+    "chromosome",
+    "nucleic acid",
+  ],
   Divide: ["calculate", "maths", "division", "operator", "code", "÷", "/"],
+  Dna: ["gene", "gmo", "helix", "heredity", "chromosome", "nucleic acid"],
   Dock: [
     "desktop",
     "applications",
@@ -2977,18 +2813,8 @@ export const TAGS: Record<string, string[]> = {
     "macos",
     "osx",
   ],
-  DnaOff: [
-    "gene",
-    "gmo free",
-    "helix",
-    "heredity",
-    "chromosome",
-    "nucleic acid",
-  ],
-  Dna: ["gene", "gmo", "helix", "heredity", "chromosome", "nucleic acid"],
   Dog: ["animal", "pet", "puppy", "hound", "canine"],
-  DollarSign: ["currency", "money", "payment"],
-  BotMessageSquare: ["robot", "ai", "chat", "assistant"],
+  Clock12: ["time", "watch", "alarm", "noon", "midnight"],
   Donut: [
     "doughnut",
     "sprinkles",
@@ -3003,6 +2829,7 @@ export const TAGS: Record<string, string[]> = {
     "hollow",
     "ring",
   ],
+  DollarSign: ["currency", "money", "payment"],
   DoorClosed: [
     "entrance",
     "entry",
@@ -3013,17 +2840,7 @@ export const TAGS: Record<string, string[]> = {
     "gateway",
     "emergency exit",
   ],
-  DoorOpen: [
-    "entrance",
-    "entry",
-    "exit",
-    "ingress",
-    "egress",
-    "gate",
-    "gateway",
-    "emergency exit",
-  ],
-  Download: ["import", "export", "save"],
+  BotMessageSquare: ["robot", "ai", "chat", "assistant"],
   Dot: [
     "interpunct",
     "interpoint",
@@ -3042,6 +2859,17 @@ export const TAGS: Record<string, string[]> = {
     "center",
     ".",
   ],
+  DoorOpen: [
+    "entrance",
+    "entry",
+    "exit",
+    "ingress",
+    "egress",
+    "gate",
+    "gateway",
+    "emergency exit",
+  ],
+  Download: ["import", "export", "save"],
   DraftingCompass: [
     "geometry",
     "trigonometry",
@@ -3055,7 +2883,18 @@ export const TAGS: Record<string, string[]> = {
     "draw",
     "sketch",
   ],
+  Drama: ["drama", "masks", "theater", "theatre", "entertainment", "show"],
   Dribbble: ["design", "social"],
+  Drill: [
+    "power",
+    "bit",
+    "head",
+    "hole",
+    "diy",
+    "toolbox",
+    "build",
+    "construction",
+  ],
   Droplet: [
     "water",
     "weather",
@@ -3067,29 +2906,6 @@ export const TAGS: Record<string, string[]> = {
     "bead",
     "globule",
   ],
-  Drama: ["drama", "masks", "theater", "theatre", "entertainment", "show"],
-  Drill: [
-    "power",
-    "bit",
-    "head",
-    "hole",
-    "diy",
-    "toolbox",
-    "build",
-    "construction",
-  ],
-  Droplets: [
-    "water",
-    "weather",
-    "liquid",
-    "fluid",
-    "wet",
-    "moisture",
-    "damp",
-    "bead",
-    "globule",
-  ],
-  Dumbbell: ["barbell", "weight", "workout", "gym"],
   Drum: [
     "drummer",
     "kit",
@@ -3111,6 +2927,18 @@ export const TAGS: Record<string, string[]> = {
     "loud",
   ],
   Drumstick: ["food", "chicken", "meat"],
+  Droplets: [
+    "water",
+    "weather",
+    "liquid",
+    "fluid",
+    "wet",
+    "moisture",
+    "damp",
+    "bead",
+    "globule",
+  ],
+  Dumbbell: ["barbell", "weight", "workout", "gym"],
   Ear: ["hearing", "noise", "audio", "accessibility"],
   EarOff: [
     "hearing",
@@ -3134,7 +2962,6 @@ export const TAGS: Record<string, string[]> = {
     "protection",
     "connection",
   ],
-  EggFried: ["food", "breakfast"],
   Earth: ["world", "browser", "language", "translate", "globe"],
   Eclipse: [
     "lunar",
@@ -3161,7 +2988,9 @@ export const TAGS: Record<string, string[]> = {
     "blend",
     "shade",
   ],
+  EggFried: ["food", "breakfast"],
   EggOff: ["egg free", "vegan", "hatched", "bad egg"],
+  EllipsisVertical: ["menu", "options", "..."],
   Egg: [
     "bird",
     "chicken",
@@ -3195,15 +3024,14 @@ export const TAGS: Record<string, string[]> = {
     "...",
   ],
   EqualNot: ["calculate", "off", "maths", "operator", "code", "≠"],
-  EllipsisVertical: ["menu", "options", "..."],
-  Eraser: ["pencil", "drawing", "undo", "delete", "clear"],
   Equal: ["calculate", "maths", "operator", "assignment", "code", "="],
+  Eraser: ["pencil", "drawing", "undo", "delete", "clear"],
   Euro: ["currency", "money", "payment"],
-  ExternalLink: ["outbound", "open", "share"],
-  EyeOff: ["view", "watch", "hide", "hidden"],
   Expand: ["scale", "fullscreen"],
-  Eye: ["view", "watch"],
+  EyeOff: ["view", "watch", "hide", "hidden"],
+  ExternalLink: ["outbound", "open", "share"],
   Facebook: ["logo", "social"],
+  Eye: ["view", "watch"],
   Factory: [
     "building",
     "business",
@@ -3212,7 +3040,6 @@ export const TAGS: Record<string, string[]> = {
     "manufacture",
     "sector",
   ],
-  FastForward: ["music"],
   Fan: ["air", "cooler", "ventilation", "ventilator", "blower"],
   Feather: ["logo"],
   Fence: [
@@ -3226,7 +3053,7 @@ export const TAGS: Record<string, string[]> = {
     "property",
     "territory",
   ],
-  Figma: ["logo", "design", "tool"],
+  FastForward: ["music"],
   FerrisWheel: [
     "big wheel",
     "daisy wheel",
@@ -3238,10 +3065,10 @@ export const TAGS: Record<string, string[]> = {
     "funfair",
   ],
   FileArchive: ["zip", "package", "archive"],
-  FileAudio2: ["music", "audio", "sound", "headphones"],
-  FileAxis3d: ["model", "3d", "axis", "coordinates"],
   FileAudio: ["music", "audio", "sound", "headphones"],
-  FileBadge: ["award", "achievement", "badge", "rosette", "prize", "winner"],
+  Figma: ["logo", "design", "tool"],
+  FileAxis3d: ["model", "3d", "axis", "coordinates"],
+  FileAudio2: ["music", "audio", "sound", "headphones"],
   FileBadge2: ["award", "achievement", "badge", "rosette", "prize", "winner"],
   FileBarChart2: [
     "statistics",
@@ -3250,13 +3077,13 @@ export const TAGS: Record<string, string[]> = {
     "graph",
     "presentation",
   ],
+  FileBadge: ["award", "achievement", "badge", "rosette", "prize", "winner"],
   FileBarChart: ["statistics", "analytics", "diagram", "graph", "presentation"],
   FileBox: ["box", "package", "model"],
+  FileCheck: ["done", "document", "todo", "tick", "complete", "task"],
   FileCheck2: ["done", "document", "todo", "tick", "complete", "task"],
   FileClock: ["history", "log", "clock"],
   FileCode2: ["script", "document", "html", "xml", "property list", "plist"],
-  FileCheck: ["done", "document", "todo", "tick", "complete", "task"],
-  FileCog: ["executable", "settings", "cog", "edit", "gear"],
   FileCode: [
     "script",
     "document",
@@ -3266,15 +3093,17 @@ export const TAGS: Record<string, string[]> = {
     "property list",
     "plist",
   ],
+  FileCog: ["executable", "settings", "cog", "edit", "gear"],
   FileDiff: ["diff", "patch"],
   FileDigit: ["number", "document"],
   FileDown: ["download", "import", "export"],
   FileHeart: ["heart", "favourite", "bookmark", "quick link"],
-  FileImage: ["image", "graphics", "photo", "picture"],
   FileInput: ["document"],
-  FileKey2: ["key", "private", "public", "security"],
-  FileJson: ["code", "json", "curly braces", "curly brackets"],
+  FileImage: ["image", "graphics", "photo", "picture"],
   FileJson2: ["code", "json", "curly braces", "curly brackets"],
+  FileJson: ["code", "json", "curly braces", "curly brackets"],
+  FileKey: ["key", "private", "public", "security"],
+  FileKey2: ["key", "private", "public", "security"],
   FileLineChart: [
     "statistics",
     "analytics",
@@ -3282,9 +3111,8 @@ export const TAGS: Record<string, string[]> = {
     "graph",
     "presentation",
   ],
-  FileKey: ["key", "private", "public", "security"],
-  FileLock: ["lock", "password", "security"],
   FileLock2: ["lock", "password", "security"],
+  FileLock: ["lock", "password", "security"],
   FileMinus2: ["document"],
   FileMinus: ["delete", "remove", "erase", "document"],
   FileOutput: ["document"],
@@ -3304,14 +3132,15 @@ export const TAGS: Record<string, string[]> = {
     "midi",
     "octave",
   ],
-  FilePenLine: ["edit"],
   FilePieChart: ["statistics", "analytics", "diagram", "graph", "presentation"],
+  FilePenLine: ["edit"],
   FilePen: ["signature"],
-  FilePlus2: ["add", "create", "new", "document"],
   FilePlus: ["add", "create", "new", "document"],
+  FilePlus2: ["add", "create", "new", "document"],
   FileQuestion: ["readme", "help", "question"],
   FileSearch2: ["lost", "document", "find", "browser"],
   FileScan: ["scan", "code", "qr-code"],
+  FileSearch: ["lost", "document", "find", "browser"],
   FileSliders: [
     "cogged",
     "gear",
@@ -3326,7 +3155,6 @@ export const TAGS: Record<string, string[]> = {
     "edit",
     "executable",
   ],
-  FileSearch: ["lost", "document", "find", "browser"],
   FileSpreadsheet: ["spreadsheet", "sheet", "table"],
   FileStack: [
     "versions",
@@ -3341,13 +3169,13 @@ export const TAGS: Record<string, string[]> = {
   FileTerminal: ["terminal", "bash", "script", "executable"],
   FileText: ["data", "txt", "pdf", "document"],
   FileType2: ["font", "text", "typography", "type"],
-  FileUp: ["upload", "import", "export"],
-  FileVideo: ["movie", "video", "film"],
   FileType: ["font", "text", "typography", "type"],
+  FileUp: ["upload", "import", "export"],
   FileVideo2: ["movie", "video", "film"],
+  FileVideo: ["movie", "video", "film"],
   FileVolume: ["audio", "music", "volume"],
-  FileX2: ["lost", "delete", "remove", "document"],
   FileVolume2: ["audio", "music", "volume"],
+  FileX2: ["lost", "delete", "remove", "document"],
   FileWarning: [
     "hidden",
     "warning",
@@ -3356,11 +3184,11 @@ export const TAGS: Record<string, string[]> = {
     "protected",
     "exclamation mark",
   ],
-  File: ["document"],
-  Files: ["multiple", "copy", "documents"],
   FileX: ["lost", "delete", "remove", "document"],
-  Film: ["movie", "video", "reel", "camera", "cinema", "entertainment"],
+  Files: ["multiple", "copy", "documents"],
+  File: ["document"],
   FilterX: ["funnel", "hopper"],
+  Film: ["movie", "video", "reel", "camera", "cinema", "entertainment"],
   Filter: ["funnel", "hopper"],
   Fingerprint: ["2fa", "authentication", "biometric", "identity", "security"],
   FireExtinguisher: [
@@ -3410,6 +3238,7 @@ export const TAGS: Record<string, string[]> = {
     "marine",
   ],
   FlagOff: ["unflag"],
+  FlagTriangleLeft: ["report", "timeline"],
   Fish: [
     "dish",
     "restaurant",
@@ -3420,7 +3249,6 @@ export const TAGS: Record<string, string[]> = {
     "sea",
     "marine",
   ],
-  FlagTriangleLeft: ["report", "timeline"],
   FlagTriangleRight: ["report", "timeline"],
   Flag: ["report"],
   FlameKindling: [
@@ -3435,16 +3263,6 @@ export const TAGS: Record<string, string[]> = {
     "sticks",
   ],
   FlashlightOff: ["torch"],
-  FlaskConicalOff: [
-    "beaker",
-    "erlenmeyer",
-    "non toxic",
-    "lab",
-    "chemistry",
-    "experiment",
-    "test",
-  ],
-  Flashlight: ["torch"],
   Flame: [
     "fire",
     "lit",
@@ -3459,6 +3277,16 @@ export const TAGS: Record<string, string[]> = {
     "station",
     "emergency",
   ],
+  Flashlight: ["torch"],
+  FlaskConicalOff: [
+    "beaker",
+    "erlenmeyer",
+    "non toxic",
+    "lab",
+    "chemistry",
+    "experiment",
+    "test",
+  ],
   FlaskConical: [
     "beaker",
     "erlenmeyer",
@@ -3467,20 +3295,19 @@ export const TAGS: Record<string, string[]> = {
     "experiment",
     "test",
   ],
-  FlaskRound: ["beaker", "lab", "chemistry", "experiment", "test"],
   FlipHorizontal2: ["reflect", "mirror", "alignment", "dashed"],
-  FlipHorizontal: ["reflect", "mirror", "alignment", "dashed"],
+  FlaskRound: ["beaker", "lab", "chemistry", "experiment", "test"],
   FlipVertical2: ["reflect", "mirror", "alignment", "dashed"],
-  FlipVertical: ["reflect", "mirror", "alignment", "dashed"],
+  FlipHorizontal: ["reflect", "mirror", "alignment", "dashed"],
   Flower2: ["sustainability", "nature", "plant"],
+  FlipVertical: ["reflect", "mirror", "alignment", "dashed"],
   Flower: ["sustainability", "nature", "plant", "spring"],
   Focus: ["camera", "lens", "photo", "dashed"],
   FoldHorizontal: ["arrow", "collapse", "fold", "vertical", "dashed"],
   FoldVertical: ["arrow", "collapse", "fold", "vertical", "dashed"],
+  FolderArchive: ["archive", "zip", "package"],
   FolderCheck: ["done", "directory", "todo", "tick", "complete", "task"],
   FolderClock: ["history", "directory", "clock"],
-  FolderArchive: ["archive", "zip", "package"],
-  FolderCode: ["directory", "coding", "develop", "software"],
   FolderClosed: ["directory", "closed"],
   FolderCog: [
     "directory",
@@ -3510,10 +3337,11 @@ export const TAGS: Record<string, string[]> = {
     "admin",
     "edit",
   ],
-  FolderGit2: ["directory", "root", "project", "git", "repo"],
-  FolderHeart: ["directory", "heart", "favourite", "bookmark", "quick link"],
-  FolderGit: ["directory", "root", "project", "git", "repo"],
   FolderDown: ["directory", "download", "import", "export"],
+  FolderGit2: ["directory", "root", "project", "git", "repo"],
+  FolderGit: ["directory", "root", "project", "git", "repo"],
+  FolderInput: ["directory", "import", "export"],
+  FolderHeart: ["directory", "heart", "favourite", "bookmark", "quick link"],
   FolderKanban: [
     "projects",
     "manage",
@@ -3534,7 +3362,6 @@ export const TAGS: Record<string, string[]> = {
     "root",
   ],
   FolderKey: ["directory", "key", "private", "security", "protected"],
-  FolderInput: ["directory", "import", "export"],
   FolderLock: ["directory", "lock", "private", "security", "protected"],
   FolderMinus: ["directory", "remove", "delete"],
   FolderOpenDot: [
@@ -3546,13 +3373,13 @@ export const TAGS: Record<string, string[]> = {
     "pinned",
   ],
   FolderOpen: ["directory"],
-  FolderOutput: ["directory", "import", "export"],
-  FolderPlus: ["directory", "add", "create", "new"],
   FolderPen: ["directory", "rename"],
   FolderRoot: ["directory", "root", "project", "git", "repo"],
+  FolderOutput: ["directory", "import", "export"],
+  FolderPlus: ["directory", "add", "create", "new"],
+  FolderSearch: ["directory", "search", "find", "lost", "browser"],
   FolderSearch2: ["directory", "search", "find", "lost", "browser"],
   FolderSymlink: ["directory", "symlink", "symbolic", "link"],
-  FolderSearch: ["directory", "search", "find", "lost", "browser"],
   FolderSync: [
     "directory",
     "synchronize",
@@ -3562,19 +3389,41 @@ export const TAGS: Record<string, string[]> = {
     "transfer",
     "backup",
   ],
+  FolderUp: ["directory", "upload", "import", "export"],
   FolderTree: ["directory", "tree", "browser"],
   FolderX: ["directory", "remove", "delete"],
-  FolderUp: ["directory", "upload", "import", "export"],
-  Folder: ["directory"],
-  Folders: ["multiple", "copy", "directories"],
-  Forward: ["send", "share", "email"],
-  Forklift: ["vehicle", "transport", "logistics"],
   Footprints: ["steps", "walking", "foot", "feet", "trail", "shoe"],
+  Folder: ["directory"],
+  Forklift: ["vehicle", "transport", "logistics"],
+  Folders: ["multiple", "copy", "directories"],
+  Framer: ["logo", "design", "tool"],
+  Forward: ["send", "share", "email"],
   Frame: ["logo", "design", "tool"],
+  Frown: ["emoji", "face", "bad", "sad", "emotion"],
   Fuel: ["filling-station", "gas", "petrol", "tank"],
   Fullscreen: ["expand", "zoom", "preview", "focus", "camera", "lens", "image"],
-  Framer: ["logo", "design", "tool"],
   GalleryHorizontalEnd: [
+    "carousel",
+    "pictures",
+    "images",
+    "scroll",
+    "swipe",
+    "album",
+    "portfolio",
+    "history",
+    "versions",
+    "backup",
+    "time machine",
+  ],
+  GalleryThumbnails: [
+    "carousel",
+    "pictures",
+    "images",
+    "album",
+    "portfolio",
+    "preview",
+  ],
+  GalleryVerticalEnd: [
     "carousel",
     "pictures",
     "images",
@@ -3596,15 +3445,6 @@ export const TAGS: Record<string, string[]> = {
     "album",
     "portfolio",
   ],
-  GalleryThumbnails: [
-    "carousel",
-    "pictures",
-    "images",
-    "album",
-    "portfolio",
-    "preview",
-  ],
-  Frown: ["emoji", "face", "bad", "sad", "emotion"],
   GalleryVertical: [
     "carousel",
     "pictures",
@@ -3616,6 +3456,15 @@ export const TAGS: Record<string, string[]> = {
   ],
   Gamepad2: ["console"],
   Gamepad: ["console"],
+  Gauge: [
+    "dashboard",
+    "dial",
+    "meter",
+    "speed",
+    "pressure",
+    "measure",
+    "level",
+  ],
   GanttChart: [
     "projects",
     "manage",
@@ -3635,17 +3484,6 @@ export const TAGS: Record<string, string[]> = {
     "code",
     "coding",
   ],
-  Gauge: [
-    "dashboard",
-    "dial",
-    "meter",
-    "speed",
-    "pressure",
-    "measure",
-    "level",
-  ],
-  Gavel: ["hammer", "mallet"],
-  Ghost: ["pac-man", "spooky"],
   Gem: [
     "diamond",
     "crystal",
@@ -3661,9 +3499,11 @@ export const TAGS: Record<string, string[]> = {
     "marriage",
     "rubygems",
   ],
+  Gavel: ["hammer", "mallet"],
   Gift: ["present", "box", "birthday", "party"],
+  Ghost: ["pac-man", "spooky"],
   GitBranchPlus: ["add", "create", "+"],
-  GitCommitVertical: ["code", "version control", "waypoint", "stop", "station"],
+  GitBranch: ["code", "version control"],
   GitCommitHorizontal: [
     "code",
     "version control",
@@ -3671,10 +3511,10 @@ export const TAGS: Record<string, string[]> = {
     "stop",
     "station",
   ],
-  GitBranch: ["code", "version control"],
+  GitCommitVertical: ["code", "version control", "waypoint", "stop", "station"],
   GitCompareArrows: ["code", "version control", "diff"],
-  GitFork: ["code", "version control"],
   GitCompare: ["code", "version control", "diff"],
+  GitFork: ["code", "version control"],
   GitGraph: ["code", "version control", "commit graph", "commits", "gitlens"],
   GitMerge: ["code", "version control"],
   GitPullRequestArrow: ["code", "version control", "open"],
@@ -3694,12 +3534,13 @@ export const TAGS: Record<string, string[]> = {
     "add",
     "+",
   ],
-  GitPullRequestDraft: ["code", "version control", "open", "draft", "dashed"],
-  GitPullRequest: ["code", "version control", "open"],
   GitPullRequestCreate: ["code", "version control", "open", "plus", "add", "+"],
   Github: ["logo", "version control"],
-  GlassWater: ["beverage", "drink", "glass", "water"],
+  GitPullRequestDraft: ["code", "version control", "open", "draft", "dashed"],
+  GitPullRequest: ["code", "version control", "open"],
   Gitlab: ["logo", "version control"],
+  GlassWater: ["beverage", "drink", "glass", "water"],
+  Glasses: ["glasses", "spectacles"],
   GlobeLock: [
     "vpn",
     "private",
@@ -3712,9 +3553,10 @@ export const TAGS: Record<string, string[]> = {
     "protection",
     "connection",
   ],
-  Glasses: ["glasses", "spectacles"],
   Globe: ["world", "browser", "language", "translate"],
   Goal: ["flag", "bullseye"],
+  Grab: ["hand"],
+  Grape: ["fruit", "wine", "food"],
   GraduationCap: [
     "school",
     "university",
@@ -3730,8 +3572,6 @@ export const TAGS: Record<string, string[]> = {
     "master's",
     "doctorate",
   ],
-  Grab: ["hand"],
-  Grape: ["fruit", "wine", "food"],
   Grid2x2Check: [
     "table",
     "rows",
@@ -3742,23 +3582,6 @@ export const TAGS: Record<string, string[]> = {
     "geometry",
     "measure",
     "data",
-    "size",
-    "width",
-    "height",
-    "distance",
-    "surface area",
-    "square meter",
-    "acre",
-  ],
-  Grid2x2: [
-    "table",
-    "rows",
-    "columns",
-    "blocks",
-    "plot",
-    "land",
-    "geometry",
-    "measure",
     "size",
     "width",
     "height",
@@ -3785,31 +3608,27 @@ export const TAGS: Record<string, string[]> = {
     "square meter",
     "acre",
   ],
-  Grid3x3: ["table", "rows", "columns"],
+  Grid2x2: [
+    "table",
+    "rows",
+    "columns",
+    "blocks",
+    "plot",
+    "land",
+    "geometry",
+    "measure",
+    "size",
+    "width",
+    "height",
+    "distance",
+    "surface area",
+    "square meter",
+    "acre",
+  ],
   GripHorizontal: ["grab", "dots", "handle", "move", "drag"],
-  Group: [
-    "cubes",
-    "packages",
-    "parts",
-    "units",
-    "collection",
-    "cluster",
-    "gather",
-    "dashed",
-  ],
-  GripVertical: ["grab", "dots", "handle", "move", "drag"],
   Grip: ["grab", "dots", "handle", "move", "drag"],
-  Ham: [
-    "food",
-    "pork",
-    "pig",
-    "meat",
-    "bone",
-    "hock",
-    "knuckle",
-    "gammon",
-    "cured",
-  ],
+  Grid3x3: ["table", "rows", "columns"],
+  GripVertical: ["grab", "dots", "handle", "move", "drag"],
   Guitar: [
     "acoustic",
     "instrument",
@@ -3827,6 +3646,27 @@ export const TAGS: Record<string, string[]> = {
     "audio",
     "sound",
     "noise",
+  ],
+  Group: [
+    "cubes",
+    "packages",
+    "parts",
+    "units",
+    "collection",
+    "cluster",
+    "gather",
+    "dashed",
+  ],
+  Ham: [
+    "food",
+    "pork",
+    "pig",
+    "meat",
+    "bone",
+    "hock",
+    "knuckle",
+    "gammon",
+    "cured",
   ],
   Hammer: ["mallet", "nails", "diy", "toolbox", "build", "construction"],
   HandCoins: [
@@ -3846,8 +3686,9 @@ export const TAGS: Record<string, string[]> = {
     "pennies",
   ],
   HandHeart: ["love", "like", "emotion"],
-  HandMetal: ["rock"],
   HandHelping: ["agreement", "help", "proposal", "charity", "begging", "terms"],
+  HandMetal: ["rock"],
+  Hand: ["wave", "move", "mouse", "grab"],
   HandPlatter: [
     "waiter",
     "waitress",
@@ -3860,7 +3701,17 @@ export const TAGS: Record<string, string[]> = {
     "course",
     "luxury",
   ],
-  Hand: ["wave", "move", "mouse", "grab"],
+  Handshake: [
+    "agreement",
+    "partnership",
+    "deal",
+    "business",
+    "assistance",
+    "cooperation",
+    "friendship",
+    "union",
+    "terms",
+  ],
   HardDriveDownload: [
     "computer",
     "server",
@@ -3881,18 +3732,6 @@ export const TAGS: Record<string, string[]> = {
     "hard disk",
     "save",
   ],
-  Handshake: [
-    "agreement",
-    "partnership",
-    "deal",
-    "business",
-    "assistance",
-    "cooperation",
-    "friendship",
-    "union",
-    "terms",
-  ],
-  Hash: ["hashtag", "number", "pound"],
   HardDrive: [
     "computer",
     "server",
@@ -3903,7 +3742,9 @@ export const TAGS: Record<string, string[]> = {
     "hard disk",
   ],
   HardHat: ["helmet", "construction", "safety", "savety"],
+  Hash: ["hashtag", "number", "pound"],
   Haze: ["mist", "fog"],
+  Heading1: ["h1", "html", "markup", "markdown"],
   HdmiPort: [
     "socket",
     "plug",
@@ -3921,15 +3762,13 @@ export const TAGS: Record<string, string[]> = {
     "input",
     "output",
   ],
-  Heading1: ["h1", "html", "markup", "markdown"],
   Heading2: ["h2", "html", "markup", "markdown"],
+  Heading3: ["h3", "html", "markup", "markdown"],
   Heading4: ["h4", "html", "markup", "markdown"],
   Heading5: ["h5", "html", "markup", "markdown"],
-  Heading3: ["h3", "html", "markup", "markdown"],
-  Headphones: ["music", "audio", "sound"],
   Heading6: ["h6", "html", "markup", "markdown"],
   Heading: ["h1", "html", "markup", "markdown"],
-  HeartCrack: ["heartbreak", "sadness", "emotion"],
+  Headphones: ["music", "audio", "sound"],
   Headset: [
     "music",
     "audio",
@@ -3944,6 +3783,8 @@ export const TAGS: Record<string, string[]> = {
     "voip",
     "video",
   ],
+  HeartCrack: ["heartbreak", "sadness", "emotion"],
+  HeartOff: ["unlike", "dislike", "hate", "emotion"],
   HeartHandshake: [
     "agreement",
     "charity",
@@ -3954,7 +3795,6 @@ export const TAGS: Record<string, string[]> = {
     "together",
     "handshake",
   ],
-  HeartOff: ["unlike", "dislike", "hate", "emotion"],
   HeartPulse: [
     "heartbeat",
     "pulse",
@@ -3977,16 +3817,6 @@ export const TAGS: Record<string, string[]> = {
     "amenities",
   ],
   Hexagon: ["shape", "node.js", "logo"],
-  HopOff: [
-    "beer",
-    "brewery",
-    "drink",
-    "hop free",
-    "allergy",
-    "intolerance",
-    "diet",
-  ],
-  Highlighter: ["mark", "text"],
   History: [
     "time",
     "redo",
@@ -3999,6 +3829,16 @@ export const TAGS: Record<string, string[]> = {
     "rotate",
     "ccw",
   ],
+  HopOff: [
+    "beer",
+    "brewery",
+    "drink",
+    "hop free",
+    "allergy",
+    "intolerance",
+    "diet",
+  ],
+  Hop: ["beer", "brewery", "drink"],
   Hospital: [
     "infirmary",
     "sanatorium",
@@ -4012,8 +3852,8 @@ export const TAGS: Record<string, string[]> = {
     "medical",
     "vet",
   ],
-  Hop: ["beer", "brewery", "drink"],
   Hotel: ["building", "hostel", "motel", "inn"],
+  Highlighter: ["mark", "text"],
   Hourglass: ["timer", "time", "sandglass"],
   HousePlug: [
     "home",
@@ -4034,6 +3874,7 @@ export const TAGS: Record<string, string[]> = {
     "residence",
     "architecture",
   ],
+  House: ["home", "living", "building", "residence", "architecture"],
   IceCreamBowl: [
     "gelato",
     "food",
@@ -4043,14 +3884,13 @@ export const TAGS: Record<string, string[]> = {
     "course",
     "meal",
   ],
-  House: ["home", "living", "building", "residence", "architecture"],
-  IceCreamCone: ["gelato", "food"],
   ImageDown: ["picture", "photo", "download", "save", "export"],
+  IceCreamCone: ["gelato", "food"],
   ImageMinus: ["remove", "delete"],
   ImageOff: ["picture", "photo"],
-  ImagePlus: ["add", "create"],
   ImagePlay: ["picture", "gif", "photo"],
   ImageUp: ["picture", "photo", "upload", "import"],
+  ImagePlus: ["add", "create"],
   Image: ["picture", "photo"],
   Images: [
     "picture",
@@ -4065,19 +3905,18 @@ export const TAGS: Record<string, string[]> = {
   ],
   Import: ["save"],
   Inbox: ["email"],
-  IndentIncrease: ["text", "tab"],
-  IndianRupee: ["currency", "money", "payment"],
   IndentDecrease: ["text", "tab"],
+  IndianRupee: ["currency", "money", "payment"],
+  IndentIncrease: ["text", "tab"],
+  Info: ["help"],
   Infinity: ["unlimited", "forever", "loop", "maths"],
   InspectionPanel: ["access", "cover", "tile", "metal", "materials", "screws"],
-  Info: ["help"],
+  Italic: ["oblique", "text", "format"],
   Instagram: ["logo", "camera", "social"],
   IterationCcw: ["arrow", "right"],
-  Italic: ["oblique", "text", "format"],
   IterationCw: ["arrow", "left"],
   JapaneseYen: ["currency", "money", "payment"],
   Joystick: ["game", "console", "control stick"],
-  KeyRound: ["password", "login", "authentication", "secure", "unlock"],
   Kanban: [
     "projects",
     "manage",
@@ -4093,6 +3932,17 @@ export const TAGS: Record<string, string[]> = {
     "agile",
     "code",
     "coding",
+  ],
+  KeyRound: ["password", "login", "authentication", "secure", "unlock"],
+  Key: [
+    "password",
+    "login",
+    "authentication",
+    "secure",
+    "unlock",
+    "keychain",
+    "key ring",
+    "fob",
   ],
   KeySquare: [
     "password",
@@ -4126,25 +3976,12 @@ export const TAGS: Record<string, string[]> = {
     "performance",
     "concert",
   ],
-  Key: [
-    "password",
-    "login",
-    "authentication",
-    "secure",
-    "unlock",
-    "keychain",
-    "key ring",
-    "fob",
-  ],
   KeyboardOff: ["unkeys", "layout", "spell", "settings", "mouse"],
-  LampCeiling: ["lighting", "household", "home", "furniture"],
   Keyboard: ["layout", "spell", "settings", "mouse"],
-  LampFloor: ["lighting", "household", "floor", "home", "furniture"],
+  LampCeiling: ["lighting", "household", "home", "furniture"],
   LampDesk: ["lighting", "household", "office", "desk", "home", "furniture"],
+  LampFloor: ["lighting", "household", "floor", "home", "furniture"],
   LampWallDown: ["lighting", "household", "wall", "home", "furniture"],
-  LampWallUp: ["lighting", "household", "wall", "home", "furniture"],
-  Lamp: ["lighting", "household", "home", "furniture"],
-  Languages: ["translate"],
   LandPlot: [
     "area",
     "surface",
@@ -4161,23 +3998,14 @@ export const TAGS: Record<string, string[]> = {
     "golf course",
     "hole",
   ],
+  Lamp: ["lighting", "household", "home", "furniture"],
+  LampWallUp: ["lighting", "household", "wall", "home", "furniture"],
+  LaptopMinimal: ["computer", "screen", "remote"],
+  Languages: ["translate"],
   Landmark: ["bank", "building", "capitol", "finance", "money"],
   Laptop: ["computer", "screen", "remote"],
-  LaptopMinimal: ["computer", "screen", "remote"],
-  Lasso: ["select", "cursor"],
   LassoSelect: ["select", "cursor"],
-  Laugh: ["emoji", "face", "happy", "good", "emotion"],
-  Layers3: [
-    "stack",
-    "pile",
-    "pages",
-    "sheets",
-    "paperwork",
-    "copies",
-    "copy",
-    "duplicate",
-    "triple",
-  ],
+  Lasso: ["select", "cursor"],
   Layers: ["stack", "pile", "pages", "sheets", "paperwork", "copies", "copy"],
   Layers2: [
     "stack",
@@ -4190,6 +4018,18 @@ export const TAGS: Record<string, string[]> = {
     "duplicate",
     "double",
     "shortcuts",
+  ],
+  Laugh: ["emoji", "face", "happy", "good", "emotion"],
+  Layers3: [
+    "stack",
+    "pile",
+    "pages",
+    "sheets",
+    "paperwork",
+    "copies",
+    "copy",
+    "duplicate",
+    "triple",
   ],
   LayoutDashboard: ["masonry", "brick"],
   LayoutList: ["todo", "tasks", "items", "pending", "image", "photo"],
@@ -4205,9 +4045,9 @@ export const TAGS: Record<string, string[]> = {
     "structure",
   ],
   LayoutTemplate: ["window", "webpage", "block", "section"],
-  Lectern: ["pulpit", "podium", "stand"],
   LeafyGreen: ["salad", "lettuce", "vegetable", "chard", "cabbage", "bok choy"],
   Leaf: ["sustainability", "nature", "energy", "plant", "autumn"],
+  Lectern: ["pulpit", "podium", "stand"],
   LibraryBig: [
     "books",
     "reading",
@@ -4258,6 +4098,8 @@ export const TAGS: Record<string, string[]> = {
     "music",
     "package",
   ],
+  Ligature: ["text", "font", "typography", "alternates", "alternatives"],
+  LightbulbOff: ["lights"],
   LifeBuoy: [
     "preserver",
     "life belt",
@@ -4271,13 +4113,11 @@ export const TAGS: Record<string, string[]> = {
     "wheel",
     "donut",
   ],
-  Ligature: ["text", "font", "typography", "alternates", "alternatives"],
-  LightbulbOff: ["lights"],
-  LineChart: ["statistics", "analytics", "diagram", "graph"],
   Lightbulb: ["idea", "bright", "lights"],
+  LineChart: ["statistics", "analytics", "diagram", "graph"],
+  Link: ["chain", "url"],
   Link2Off: ["unchain", "chain"],
   Link2: ["chain", "url"],
-  Link: ["chain", "url"],
   Linkedin: ["logo", "social media", "social"],
   ListChecks: [
     "todo",
@@ -4289,8 +4129,6 @@ export const TAGS: Record<string, string[]> = {
     "items",
     "pending",
   ],
-  ListEnd: ["queue", "bottom", "end", "playlist"],
-  ListCheck: ["done", "check", "tick", "complete", "list", "to-do", "bom"],
   ListCollapse: [
     "items",
     "collapse",
@@ -4306,11 +4144,12 @@ export const TAGS: Record<string, string[]> = {
     "fold",
     "unfold",
   ],
+  ListEnd: ["queue", "bottom", "end", "playlist"],
   ListFilter: ["options"],
-  ListMinus: ["playlist", "remove", "song", "subtract", "delete", "unqueue"],
-  ListOrdered: ["number", "order", "queue"],
   ListMusic: ["playlist", "queue", "music", "audio", "playback"],
+  ListMinus: ["playlist", "remove", "song", "subtract", "delete", "unqueue"],
   ListPlus: ["playlist", "add", "song", "track", "new"],
+  ListOrdered: ["number", "order", "queue"],
   ListStart: ["queue", "top", "start", "next", "playlist"],
   ListRestart: ["reset", "refresh", "reload", "playlist", "replay"],
   ListTodo: [
@@ -4324,9 +4163,6 @@ export const TAGS: Record<string, string[]> = {
     "pending",
   ],
   ListVideo: ["playlist", "video", "playback"],
-  ListTree: ["tree", "browser"],
-  ListX: ["playlist", "subtract", "remove", "delete", "unqueue"],
-  List: ["options"],
   LoaderCircle: [
     "loading",
     "wait",
@@ -4337,6 +4173,19 @@ export const TAGS: Record<string, string[]> = {
     "throbber",
     "circle",
   ],
+  ListTree: ["tree", "browser"],
+  ListX: ["playlist", "subtract", "remove", "delete", "unqueue"],
+  List: ["options"],
+  Loader: [
+    "loading",
+    "wait",
+    "busy",
+    "progress",
+    "spinner",
+    "spinning",
+    "throbber",
+  ],
+  LocateFixed: ["map", "gps", "location", "cross"],
   LoaderPinwheel: [
     "loading",
     "wait",
@@ -4349,25 +4198,16 @@ export const TAGS: Record<string, string[]> = {
     "frozen",
     "freeze",
   ],
-  Loader: [
-    "loading",
-    "wait",
-    "busy",
-    "progress",
-    "spinner",
-    "spinning",
-    "throbber",
-  ],
-  LocateFixed: ["map", "gps", "location", "cross"],
   LocateOff: ["map", "gps", "location", "cross"],
   Locate: ["map", "gps", "location", "cross"],
-  LockKeyhole: ["security", "password", "secure", "admin"],
   LockKeyholeOpen: ["security"],
+  LockKeyhole: ["security", "password", "secure", "admin"],
   LockOpen: ["security"],
-  LogOut: ["sign out", "arrow", "exit", "auth"],
   Lock: ["security", "password", "secure", "admin"],
   LogIn: ["sign in", "arrow", "enter", "auth"],
+  LogOut: ["sign out", "arrow", "exit", "auth"],
   Logs: ["options", "list", "menu", "order", "queue", "tasks", "logs"],
+  Luggage: ["baggage", "luggage", "travel", "suitcase"],
   Lollipop: ["lolly", "candy", "sugar", "food", "sweet", "dessert", "spiral"],
   Magnet: ["horseshoe", "lock", "science", "snap"],
   MailCheck: [
@@ -4384,14 +4224,10 @@ export const TAGS: Record<string, string[]> = {
     "complete",
     "task",
   ],
-  Luggage: ["baggage", "luggage", "travel", "suitcase"],
   MailMinus: ["email", "message", "letter", "remove", "delete"],
-  MailPlus: ["email", "message", "letter", "add", "create", "new", "compose"],
   MailOpen: ["email", "message", "letter", "read"],
+  MailPlus: ["email", "message", "letter", "add", "create", "new", "compose"],
   MailQuestion: ["email", "message", "letter", "delivery", "undelivered"],
-  MailSearch: ["email", "message", "letter", "search"],
-  MailX: ["email", "message", "letter", "remove", "delete"],
-  Mail: ["email", "message", "letter", "unread"],
   MailWarning: [
     "email",
     "message",
@@ -4399,7 +4235,10 @@ export const TAGS: Record<string, string[]> = {
     "delivery error",
     "exclamation mark",
   ],
+  MailSearch: ["email", "message", "letter", "search"],
   Mailbox: ["emails", "messages", "letters", "mailing list", "newsletter"],
+  MailX: ["email", "message", "letter", "remove", "delete"],
+  Mail: ["email", "message", "letter", "unread"],
   Mails: [
     "emails",
     "messages",
@@ -4412,13 +4251,11 @@ export const TAGS: Record<string, string[]> = {
   MapPinOff: ["location", "waypoint", "marker", "remove"],
   MapPin: ["location", "waypoint", "marker", "drop"],
   MapPinned: ["location", "waypoint", "marker", "drop"],
+  Martini: ["cocktail", "alcohol", "beverage", "bar", "drink", "glass"],
   Map: ["location", "navigation", "travel"],
   Maximize2: ["fullscreen", "arrows", "expand"],
-  Martini: ["cocktail", "alcohol", "beverage", "bar", "drink", "glass"],
   Maximize: ["fullscreen", "expand", "dashed"],
   Medal: ["prize", "sports", "winner", "trophy", "award", "achievement"],
-  Megaphone: ["advertisement", "attention", "alert", "notification"],
-  Meh: ["emoji", "face", "neutral", "emotion"],
   MegaphoneOff: [
     "advertisement",
     "attention",
@@ -4427,17 +4264,7 @@ export const TAGS: Record<string, string[]> = {
     "disable",
     "silent",
   ],
-  MessageCircleCode: [
-    "comment",
-    "chat",
-    "conversation",
-    "dialog",
-    "feedback",
-    "speech bubble",
-    "code review",
-    "coding",
-  ],
-  Menu: ["bars", "navigation", "hamburger", "options"],
+  Megaphone: ["advertisement", "attention", "alert", "notification"],
   MemoryStick: [
     "ram",
     "random access",
@@ -4450,6 +4277,9 @@ export const TAGS: Record<string, string[]> = {
     "gigabytes",
     "gb",
   ],
+  Meh: ["emoji", "face", "neutral", "emotion"],
+  Menu: ["bars", "navigation", "hamburger", "options"],
+  Merge: ["combine", "join", "unite"],
   MessageCircleDashed: [
     "comment",
     "chat",
@@ -4459,7 +4289,16 @@ export const TAGS: Record<string, string[]> = {
     "speech bubble",
     "draft",
   ],
-  Merge: ["combine", "join", "unite"],
+  MessageCircleCode: [
+    "comment",
+    "chat",
+    "conversation",
+    "dialog",
+    "feedback",
+    "speech bubble",
+    "code review",
+    "coding",
+  ],
   MessageCircleHeart: [
     "comment",
     "chat",
@@ -4474,6 +4313,22 @@ export const TAGS: Record<string, string[]> = {
     "dating",
     "date",
     "speech bubble",
+  ],
+  MessageCircleOff: [
+    "comment",
+    "chat",
+    "conversation",
+    "dialog",
+    "feedback",
+    "speech bubble",
+    "clear",
+    "close",
+    "delete",
+    "remove",
+    "cancel",
+    "silence",
+    "mute",
+    "moderate",
   ],
   MessageCircleMore: [
     "comment",
@@ -4491,21 +4346,15 @@ export const TAGS: Record<string, string[]> = {
     "...",
     "…",
   ],
-  MessageCircleOff: [
+  MessageCircleReply: [
     "comment",
     "chat",
     "conversation",
     "dialog",
     "feedback",
     "speech bubble",
-    "clear",
-    "close",
-    "delete",
-    "remove",
-    "cancel",
-    "silence",
-    "mute",
-    "moderate",
+    "reply",
+    "response",
   ],
   MessageCirclePlus: [
     "comment",
@@ -4541,24 +4390,6 @@ export const TAGS: Record<string, string[]> = {
     "protected",
     "exclamation mark",
   ],
-  MessageCircleReply: [
-    "comment",
-    "chat",
-    "conversation",
-    "dialog",
-    "feedback",
-    "speech bubble",
-    "reply",
-    "response",
-  ],
-  MessageCircle: [
-    "comment",
-    "chat",
-    "conversation",
-    "dialog",
-    "feedback",
-    "speech bubble",
-  ],
   MessageCircleX: [
     "comment",
     "chat",
@@ -4575,6 +4406,14 @@ export const TAGS: Record<string, string[]> = {
     "mute",
     "moderate",
   ],
+  MessageCircle: [
+    "comment",
+    "chat",
+    "conversation",
+    "dialog",
+    "feedback",
+    "speech bubble",
+  ],
   MessageSquareCode: [
     "comment",
     "chat",
@@ -4584,15 +4423,6 @@ export const TAGS: Record<string, string[]> = {
     "speech bubble",
     "code review",
     "coding",
-  ],
-  MessageSquareDashed: [
-    "comment",
-    "chat",
-    "conversation",
-    "dialog",
-    "feedback",
-    "speech bubble",
-    "draft",
   ],
   MessageSquareDiff: [
     "comment",
@@ -4613,15 +4443,14 @@ export const TAGS: Record<string, string[]> = {
     "version control",
     "git",
   ],
-  MessageSquareDot: [
-    "unread",
-    "unresolved",
+  MessageSquareDashed: [
     "comment",
     "chat",
     "conversation",
     "dialog",
     "feedback",
     "speech bubble",
+    "draft",
   ],
   MessageSquareHeart: [
     "comment",
@@ -4636,6 +4465,16 @@ export const TAGS: Record<string, string[]> = {
     "valentine",
     "dating",
     "date",
+    "speech bubble",
+  ],
+  MessageSquareDot: [
+    "unread",
+    "unresolved",
+    "comment",
+    "chat",
+    "conversation",
+    "dialog",
+    "feedback",
     "speech bubble",
   ],
   MessageSquareMore: [
@@ -4654,28 +4493,6 @@ export const TAGS: Record<string, string[]> = {
     "...",
     "…",
   ],
-  MessageSquareQuote: [
-    "comment",
-    "chat",
-    "conversation",
-    "dialog",
-    "feedback",
-    "speech bubble",
-    "blockquote",
-    "quotation",
-    "indent",
-    "reply",
-    "response",
-  ],
-  MessageSquarePlus: [
-    "comment",
-    "chat",
-    "conversation",
-    "dialog",
-    "feedback",
-    "speech bubble",
-    "add",
-  ],
   MessageSquareOff: [
     "comment",
     "chat",
@@ -4691,6 +4508,28 @@ export const TAGS: Record<string, string[]> = {
     "silence",
     "mute",
     "moderate",
+  ],
+  MessageSquarePlus: [
+    "comment",
+    "chat",
+    "conversation",
+    "dialog",
+    "feedback",
+    "speech bubble",
+    "add",
+  ],
+  MessageSquareQuote: [
+    "comment",
+    "chat",
+    "conversation",
+    "dialog",
+    "feedback",
+    "speech bubble",
+    "blockquote",
+    "quotation",
+    "indent",
+    "reply",
+    "response",
   ],
   MessageSquareReply: [
     "comment",
@@ -4711,6 +4550,22 @@ export const TAGS: Record<string, string[]> = {
     "speech bubble",
     "network",
     "forward",
+  ],
+  MessageSquareX: [
+    "comment",
+    "chat",
+    "conversation",
+    "dialog",
+    "feedback",
+    "speech bubble",
+    "clear",
+    "close",
+    "delete",
+    "remove",
+    "cancel",
+    "silence",
+    "mute",
+    "moderate",
   ],
   MessageSquareText: [
     "comment",
@@ -4736,19 +4591,6 @@ export const TAGS: Record<string, string[]> = {
     "protected",
     "exclamation mark",
   ],
-  GalleryVerticalEnd: [
-    "carousel",
-    "pictures",
-    "images",
-    "scroll",
-    "swipe",
-    "album",
-    "portfolio",
-    "history",
-    "versions",
-    "backup",
-    "time machine",
-  ],
   MessageSquare: [
     "comment",
     "chat",
@@ -4757,6 +4599,7 @@ export const TAGS: Record<string, string[]> = {
     "feedback",
     "speech bubble",
   ],
+  MicOff: ["record", "sound", "mute", "microphone"],
   MessagesSquare: [
     "comment",
     "chat",
@@ -4767,23 +4610,6 @@ export const TAGS: Record<string, string[]> = {
     "copy",
     "multiple",
   ],
-  MessageSquareX: [
-    "comment",
-    "chat",
-    "conversation",
-    "dialog",
-    "feedback",
-    "speech bubble",
-    "clear",
-    "close",
-    "delete",
-    "remove",
-    "cancel",
-    "silence",
-    "mute",
-    "moderate",
-  ],
-  Mic: ["record", "sound", "listen", "radio", "podcast", "microphone"],
   MicVocal: [
     "lyrics",
     "voice",
@@ -4796,7 +4622,7 @@ export const TAGS: Record<string, string[]> = {
     "singing",
     "microphone",
   ],
-  MicOff: ["record", "sound", "mute", "microphone"],
+  Mic: ["record", "sound", "listen", "radio", "podcast", "microphone"],
   Microscope: ["medical", "education", "science", "imaging", "research"],
   Microwave: ["oven", "cooker", "toaster oven", "bake"],
   Milestone: [
@@ -4808,6 +4634,7 @@ export const TAGS: Record<string, string[]> = {
     "version control",
   ],
   Milk: ["lactose", "bottle", "beverage", "drink", "water", "diet"],
+  Minimize2: ["exit fullscreen", "arrows", "close", "shrink"],
   MilkOff: [
     "lactose free",
     "bottle",
@@ -4818,8 +4645,6 @@ export const TAGS: Record<string, string[]> = {
     "intolerance",
     "diet",
   ],
-  Minimize2: ["exit fullscreen", "arrows", "close", "shrink"],
-  Minimize: ["exit fullscreen", "close", "shrink"],
   MonitorCheck: [
     "tv",
     "screen",
@@ -4854,25 +4679,18 @@ export const TAGS: Record<string, string[]> = {
     "minimum",
     "downgrade",
   ],
-  MonitorCog: [
+  MonitorDown: ["tv", "screen", "display", "desktop", "download"],
+  Minimize: ["exit fullscreen", "close", "shrink"],
+  MonitorDot: [
     "tv",
     "screen",
     "display",
+    "desktop",
+    "running",
+    "active",
     "virtual machine",
     "vm",
-    "executable",
-    "settings",
-    "cog",
-    "edit",
-    "gear",
-    "configuration",
-    "preferences",
-    "system",
-    "control panel",
-    "network",
-    "computing",
   ],
-  MonitorDown: ["tv", "screen", "display", "desktop", "download"],
   MonitorOff: ["share"],
   MonitorPause: [
     "tv",
@@ -4888,15 +4706,16 @@ export const TAGS: Record<string, string[]> = {
     "virtual machine",
     "vm",
   ],
-  MonitorDot: [
-    "tv",
-    "screen",
-    "display",
+  MonitorSmartphone: [
+    "smartphone",
+    "phone",
+    "cellphone",
+    "device",
+    "mobile",
     "desktop",
-    "running",
-    "active",
-    "virtual machine",
-    "vm",
+    "monitor",
+    "responsive",
+    "screens",
   ],
   MonitorPlay: [
     "tv",
@@ -4912,18 +4731,16 @@ export const TAGS: Record<string, string[]> = {
     "virtual machine",
     "vm",
   ],
-  MonitorSmartphone: [
-    "smartphone",
-    "phone",
-    "cellphone",
-    "device",
-    "mobile",
-    "desktop",
-    "monitor",
-    "responsive",
-    "screens",
-  ],
   MonitorSpeaker: ["devices", "connect", "cast"],
+  MonitorUp: [
+    "tv",
+    "screen",
+    "display",
+    "upload",
+    "connect",
+    "remote",
+    "screen share",
+  ],
   MonitorStop: [
     "tv",
     "screen",
@@ -4936,15 +4753,6 @@ export const TAGS: Record<string, string[]> = {
     "shutdown",
     "virtual machine",
     "vm",
-  ],
-  MonitorUp: [
-    "tv",
-    "screen",
-    "display",
-    "upload",
-    "connect",
-    "remote",
-    "screen share",
   ],
   MonitorX: [
     "tv",
@@ -4959,41 +4767,64 @@ export const TAGS: Record<string, string[]> = {
     "remove",
     "delete",
   ],
-  MoonStar: ["dark", "night", "star"],
+  Monitor: ["tv", "screen", "display", "virtual machine", "vm"],
   Moon: ["dark", "night"],
+  MoonStar: ["dark", "night", "star"],
   Mountain: ["climb", "hike", "rock"],
   MountainSnow: ["alpine", "climb", "snow"],
-  Monitor: ["tv", "screen", "display", "virtual machine", "vm"],
-  MousePointer2: ["click", "select"],
   MouseOff: ["device", "scroll", "click", "disabled"],
-  MousePointerBan: ["wait", "busy", "loading", "blocked", "frozen", "freeze"],
   MousePointerClick: ["click", "select"],
-  Mouse: ["device", "scroll", "click"],
   MousePointer: ["click", "select"],
+  MousePointer2: ["click", "select"],
+  MousePointerBan: ["wait", "busy", "loading", "blocked", "frozen", "freeze"],
+  Mouse: ["device", "scroll", "click"],
   Move3d: ["arrows", "axis", "gizmo", "coordinates", "transform", "translate"],
+  MoveDiagonal2: ["double", "arrow"],
   MoveDiagonal: ["double", "arrow"],
   MoveDownLeft: ["arrow", "direction"],
-  MoveDiagonal2: ["double", "arrow"],
   MoveDownRight: ["arrow", "direction"],
   MoveDown: ["arrow", "direction", "downwards", "south"],
-  MoveLeft: ["arrow", "direction", "back", "west"],
   MoveHorizontal: ["double", "arrow"],
-  MoveRight: ["arrow", "direction", "trend flat", "east"],
+  MoveLeft: ["arrow", "direction", "back", "west"],
   MoveUpLeft: ["arrow", "direction"],
+  MoveRight: ["arrow", "direction", "trend flat", "east"],
   MoveUp: ["arrow", "direction", "upwards", "north"],
   MoveUpRight: ["arrow", "direction"],
+  MoveVertical: ["double", "arrow"],
   Move: ["arrows"],
   Music2: ["quaver", "eighth note", "note"],
-  MoveVertical: ["double", "arrow"],
+  Music3: ["crotchet", "minim", "quarter note", "half note", "note"],
   Music4: ["semiquaver", "sixteenth note", "note"],
   Music: ["note", "quaver", "eighth note"],
-  Music3: ["crotchet", "minim", "quarter note", "half note", "note"],
   Navigation2Off: ["location", "travel"],
   Navigation2: ["location", "travel"],
-  Navigation: ["location", "travel"],
   NavigationOff: ["location", "travel"],
   Network: ["tree"],
+  Newspaper: ["news", "feed", "home", "magazine", "article", "headline"],
+  Navigation: ["location", "travel"],
   Nfc: ["contactless", "payment", "near-field communication"],
+  NotebookTabs: [
+    "notepad",
+    "notes",
+    "people",
+    "family",
+    "friends",
+    "acquaintances",
+    "contacts",
+    "details",
+    "addresses",
+    "phone numbers",
+    "directory",
+    "listing",
+    "networking",
+    "alphabetical",
+    "a-z",
+    "organizer",
+    "organiser",
+    "planner",
+    "diary",
+    "stationery",
+  ],
   NotebookPen: [
     "pencil",
     "notepad",
@@ -5021,28 +4852,32 @@ export const TAGS: Record<string, string[]> = {
     "eraser",
     "rubber",
   ],
-  Newspaper: ["news", "feed", "home", "magazine", "article", "headline"],
-  NotebookTabs: [
+  NotebookText: [
     "notepad",
     "notes",
-    "people",
-    "family",
-    "friends",
-    "acquaintances",
-    "contacts",
-    "details",
-    "addresses",
-    "phone numbers",
-    "directory",
-    "listing",
-    "networking",
-    "alphabetical",
-    "a-z",
+    "pages",
+    "paper",
+    "stationery",
+    "sketchbook",
     "organizer",
     "organiser",
     "planner",
     "diary",
-    "stationery",
+    "journal",
+    "writing",
+    "write",
+    "written",
+    "reading",
+    "high school",
+    "university",
+    "college",
+    "academy",
+    "student",
+    "study",
+    "research",
+    "homework",
+    "lines",
+    "opened",
   ],
   Notebook: [
     "notepad",
@@ -5072,49 +4907,6 @@ export const TAGS: Record<string, string[]> = {
     "homework",
     "research",
   ],
-  NotepadTextDashed: [
-    "notebook",
-    "notes",
-    "pages",
-    "paper",
-    "stationery",
-    "diary",
-    "journal",
-    "writing",
-    "write",
-    "written",
-    "draft",
-    "template",
-    "lines",
-  ],
-  NotebookText: [
-    "notepad",
-    "notes",
-    "pages",
-    "paper",
-    "stationery",
-    "sketchbook",
-    "organizer",
-    "organiser",
-    "planner",
-    "diary",
-    "journal",
-    "writing",
-    "write",
-    "written",
-    "reading",
-    "high school",
-    "university",
-    "college",
-    "academy",
-    "student",
-    "study",
-    "research",
-    "homework",
-    "lines",
-    "opened",
-  ],
-  NutOff: ["hazelnut", "acorn", "food", "allergy", "intolerance", "diet"],
   NotepadText: [
     "notebook",
     "notes",
@@ -5142,14 +4934,31 @@ export const TAGS: Record<string, string[]> = {
     "lines",
     "opened",
   ],
-  Nut: ["hazelnut", "acorn", "food", "diet"],
-  OctagonX: ["delete", "stop", "alert", "warning", "times", "clear", "maths"],
+  NutOff: ["hazelnut", "acorn", "food", "allergy", "intolerance", "diet"],
+  NotepadTextDashed: [
+    "notebook",
+    "notes",
+    "pages",
+    "paper",
+    "stationery",
+    "diary",
+    "journal",
+    "writing",
+    "write",
+    "written",
+    "draft",
+    "template",
+    "lines",
+  ],
   OctagonAlert: ["warning", "alert", "danger", "exclamation mark"],
+  Nut: ["hazelnut", "acorn", "food", "diet"],
   OctagonPause: ["music", "audio", "stop"],
+  OctagonX: ["delete", "stop", "alert", "warning", "times", "clear", "maths"],
   Octagon: ["stop", "shape"],
-  Option: ["keyboard", "key", "mac", "alt", "button"],
   Orbit: ["planet", "space", "physics", "satellites", "moons"],
+  Option: ["keyboard", "key", "mac", "alt", "button"],
   Origami: ["paper", "bird"],
+  PackageMinus: ["delete", "remove"],
   Package2: [
     "box",
     "container",
@@ -5171,7 +4980,6 @@ export const TAGS: Record<string, string[]> = {
     "task",
     "delivered",
   ],
-  PackageMinus: ["delete", "remove"],
   PackageOpen: [
     "box",
     "container",
@@ -5184,7 +4992,7 @@ export const TAGS: Record<string, string[]> = {
   ],
   PackagePlus: ["new", "add", "create"],
   PackageSearch: ["find", "product process"],
-  PaintRoller: ["brush", "color", "colour", "decoration", "diy"],
+  PackageX: ["delete", "remove"],
   Package: [
     "box",
     "container",
@@ -5198,8 +5006,8 @@ export const TAGS: Record<string, string[]> = {
     "zip",
     "module",
   ],
-  PackageX: ["delete", "remove"],
   PaintBucket: ["fill", "paint", "bucket", "color", "colour"],
+  PaintRoller: ["brush", "color", "colour", "decoration", "diy"],
   PaintbrushVertical: [
     "brush",
     "paintbrush",
@@ -5209,6 +5017,16 @@ export const TAGS: Record<string, string[]> = {
     "decoration",
     "diy",
   ],
+  Paintbrush: [
+    "brush",
+    "paintbrush",
+    "design",
+    "color",
+    "colour",
+    "decoration",
+    "diy",
+  ],
+  PanelBottomClose: ["drawer", "dock", "hide", "chevron", "down"],
   Palette: [
     "colors",
     "colours",
@@ -5219,21 +5037,10 @@ export const TAGS: Record<string, string[]> = {
     "watercolour",
     "artist",
   ],
-  PanelBottomClose: ["drawer", "dock", "hide", "chevron", "down"],
-  Paintbrush: [
-    "brush",
-    "paintbrush",
-    "design",
-    "color",
-    "colour",
-    "decoration",
-    "diy",
-  ],
   PanelBottomDashed: ["drawer", "dock", "show", "reveal"],
   PanelBottomOpen: ["drawer", "dock", "show", "reveal", "chevron", "up"],
   PanelBottom: ["drawer", "dock"],
   PanelLeftClose: ["primary", "drawer", "hide", "chevron", "<"],
-  PanelLeftDashed: ["sidebar", "primary", "drawer", "show", "reveal"],
   PanelLeftOpen: [
     "primary",
     "drawer",
@@ -5243,8 +5050,12 @@ export const TAGS: Record<string, string[]> = {
     "right",
     ">",
   ],
-  PanelRightClose: ["sidebar", "secondary", "drawer", "hide", "chevron", ">"],
+  PanelLeftDashed: ["sidebar", "primary", "drawer", "show", "reveal"],
   PanelLeft: ["primary", "drawer"],
+  PanelRightClose: ["sidebar", "secondary", "drawer", "hide", "chevron", ">"],
+  PanelRightDashed: ["sidebar", "secondary", "drawer", "show", "reveal"],
+  PanelRight: ["sidebar", "secondary", "drawer"],
+  PanelTopClose: ["menu bar", "drawer", "hide", "chevron", "up"],
   PanelRightOpen: [
     "sidebar",
     "secondary",
@@ -5255,15 +5066,11 @@ export const TAGS: Record<string, string[]> = {
     "left",
     "<",
   ],
-  PanelRightDashed: ["sidebar", "secondary", "drawer", "show", "reveal"],
-  PanelTopClose: ["menu bar", "drawer", "hide", "chevron", "up"],
-  PanelRight: ["sidebar", "secondary", "drawer"],
-  PanelTopOpen: ["menu bar", "drawer", "show", "reveal", "chevron", "down"],
   PanelTopDashed: ["menu bar", "drawer", "show", "reveal"],
   PanelsLeftBottom: ["drawers", "sidebar", "primary"],
   PanelTop: ["drawer", "browser", "webpage"],
+  PanelTopOpen: ["menu bar", "drawer", "show", "reveal", "chevron", "down"],
   PanelsRightBottom: ["drawers", "sidebar", "secondary"],
-  Paperclip: ["attachment", "file"],
   PanelsTopLeft: [
     "menu bar",
     "sidebar",
@@ -5316,30 +5123,8 @@ export const TAGS: Record<string, string[]> = {
   ],
   Pause: ["music", "stop"],
   PcCase: ["computer", "chassis"],
+  Paperclip: ["attachment", "file"],
   PenLine: [
-    "pencil",
-    "change",
-    "create",
-    "draw",
-    "writer",
-    "writing",
-    "biro",
-    "ink",
-    "marker",
-    "felt tip",
-    "stationery",
-    "artist",
-  ],
-  PenTool: ["vector", "drawing", "path"],
-  PenOff: [
-    "disabled",
-    "inactive",
-    "non-editable",
-    "locked",
-    "read-only",
-    "unmodifiable",
-    "frozen",
-    "restricted",
     "pencil",
     "change",
     "create",
@@ -5367,7 +5152,7 @@ export const TAGS: Record<string, string[]> = {
     "stationery",
     "artist",
   ],
-  PencilOff: [
+  PenOff: [
     "disabled",
     "inactive",
     "non-editable",
@@ -5376,17 +5161,20 @@ export const TAGS: Record<string, string[]> = {
     "unmodifiable",
     "frozen",
     "restricted",
-    "rubber",
-    "edit",
+    "pencil",
+    "change",
     "create",
     "draw",
-    "sketch",
-    "draft",
     "writer",
     "writing",
+    "biro",
+    "ink",
+    "marker",
+    "felt tip",
     "stationery",
     "artist",
   ],
+  PenTool: ["vector", "drawing", "path"],
   PencilLine: [
     "pencil",
     "change",
@@ -5400,18 +5188,6 @@ export const TAGS: Record<string, string[]> = {
     "ink",
     "marker",
     "felt tip",
-    "stationery",
-    "artist",
-  ],
-  Pencil: [
-    "rubber",
-    "edit",
-    "create",
-    "draw",
-    "sketch",
-    "draft",
-    "writer",
-    "writing",
     "stationery",
     "artist",
   ],
@@ -5445,6 +5221,38 @@ export const TAGS: Record<string, string[]> = {
     "extent",
   ],
   Pentagon: ["shape"],
+  Pencil: [
+    "rubber",
+    "edit",
+    "create",
+    "draw",
+    "sketch",
+    "draft",
+    "writer",
+    "writing",
+    "stationery",
+    "artist",
+  ],
+  PencilOff: [
+    "disabled",
+    "inactive",
+    "non-editable",
+    "locked",
+    "read-only",
+    "unmodifiable",
+    "frozen",
+    "restricted",
+    "rubber",
+    "edit",
+    "create",
+    "draw",
+    "sketch",
+    "draft",
+    "writer",
+    "writing",
+    "stationery",
+    "artist",
+  ],
   Percent: [
     "percentage",
     "modulo",
@@ -5456,13 +5264,12 @@ export const TAGS: Record<string, string[]> = {
     "offer",
     "marketing",
   ],
-  PhoneCall: ["ring"],
   PersonStanding: ["people", "human", "accessibility", "stick figure"],
-  PhoneIncoming: ["call"],
+  PhoneCall: ["ring"],
   PhoneForwarded: ["call"],
+  PhoneIncoming: ["call"],
   PhoneMissed: ["call"],
   PhoneOff: ["call", "mute"],
-  Phone: ["call"],
   PhoneOutgoing: ["call"],
   Pi: [
     "constant",
@@ -5473,6 +5280,20 @@ export const TAGS: Record<string, string[]> = {
     "trigonometry",
     "geometry",
     "formula",
+  ],
+  Phone: ["call"],
+  Pickaxe: [
+    "mining",
+    "mine",
+    "land worker",
+    "extraction",
+    "labor",
+    "construction",
+    "progress",
+    "advancement",
+    "crafting",
+    "building",
+    "creation",
   ],
   Piano: [
     "music",
@@ -5489,19 +5310,6 @@ export const TAGS: Record<string, string[]> = {
     "pianist",
     "performance",
     "concert",
-  ],
-  Pickaxe: [
-    "mining",
-    "mine",
-    "land worker",
-    "extraction",
-    "labor",
-    "construction",
-    "progress",
-    "advancement",
-    "crafting",
-    "building",
-    "creation",
   ],
   PictureInPicture: [
     "display",
@@ -5588,9 +5396,18 @@ export const TAGS: Record<string, string[]> = {
   Pizza: ["pie", "quiche", "food"],
   PlaneLanding: ["arrival", "plane", "trip", "airplane", "landing"],
   PlaneTakeoff: ["departure", "plane", "trip", "airplane", "takeoff"],
+  Plane: ["plane", "trip", "airplane"],
   Play: ["music", "audio", "video", "start", "run"],
   Plug2: ["electricity", "energy", "socket", "outlet"],
-  Plane: ["plane", "trip", "airplane"],
+  PlugZap: [
+    "electricity",
+    "energy",
+    "electronics",
+    "charge",
+    "charging",
+    "battery",
+    "connect",
+  ],
   PlugZap2: [
     "electricity",
     "energy",
@@ -5601,25 +5418,6 @@ export const TAGS: Record<string, string[]> = {
     "connect",
   ],
   Plug: ["electricity", "energy", "electronics", "socket", "outlet"],
-  PlugZap: [
-    "electricity",
-    "energy",
-    "electronics",
-    "charge",
-    "charging",
-    "battery",
-    "connect",
-  ],
-  PocketKnife: [
-    "swiss army knife",
-    "penknife",
-    "multi-tool",
-    "multitask",
-    "blade",
-    "cutter",
-    "gadget",
-    "corkscrew",
-  ],
   Plus: [
     "add",
     "new",
@@ -5639,7 +5437,16 @@ export const TAGS: Record<string, string[]> = {
     "extra",
     "+",
   ],
-  Pocket: ["logo", "save"],
+  PocketKnife: [
+    "swiss army knife",
+    "penknife",
+    "multi-tool",
+    "multitask",
+    "blade",
+    "cutter",
+    "gadget",
+    "corkscrew",
+  ],
   Podcast: [
     "audio",
     "music",
@@ -5650,8 +5457,9 @@ export const TAGS: Record<string, string[]> = {
     "subscription",
     "stream",
   ],
-  Pointer: ["mouse"],
+  Pocket: ["logo", "save"],
   PointerOff: ["mouse"],
+  Pointer: ["mouse"],
   Popcorn: [
     "cinema",
     "movies",
@@ -5679,7 +5487,7 @@ export const TAGS: Record<string, string[]> = {
     "troubleshoot",
   ],
   PowerOff: ["on", "off", "device", "switch"],
-  Printer: ["fax", "office", "device"],
+  PrinterCheck: ["fax", "office", "device", "success", "printed"],
   Presentation: [
     "screen",
     "whiteboard",
@@ -5696,7 +5504,19 @@ export const TAGS: Record<string, string[]> = {
     "project",
     "planning",
   ],
-  PrinterCheck: ["fax", "office", "device", "success", "printed"],
+  Printer: ["fax", "office", "device"],
+  Projector: [
+    "cinema",
+    "film",
+    "movie",
+    "home video",
+    "presentation",
+    "slideshow",
+    "office",
+    "meeting",
+    "project",
+    "planning",
+  ],
   Proportions: [
     "screens",
     "sizes",
@@ -5717,18 +5537,8 @@ export const TAGS: Record<string, string[]> = {
     "portrait",
     "landscape",
   ],
-  Projector: [
-    "cinema",
-    "film",
-    "movie",
-    "home video",
-    "presentation",
-    "slideshow",
-    "office",
-    "meeting",
-    "project",
-    "planning",
-  ],
+  Puzzle: ["component", "module", "part", "piece"],
+  QrCode: ["barcode", "scan", "link", "url", "information", "digital"],
   Pyramid: [
     "prism",
     "triangle",
@@ -5741,8 +5551,6 @@ export const TAGS: Record<string, string[]> = {
     "landmark",
     "tourism",
   ],
-  Puzzle: ["component", "module", "part", "piece"],
-  QrCode: ["barcode", "scan", "link", "url", "information", "digital"],
   Quote: ["quotation"],
   Rabbit: [
     "animal",
@@ -5777,8 +5585,8 @@ export const TAGS: Record<string, string[]> = {
     "molecule",
   ],
   RadioReceiver: ["device", "music", "connect"],
-  RadioTower: ["signal", "broadcast", "connectivity", "live", "frequency"],
   Radio: ["signal", "broadcast", "connectivity", "live", "frequency"],
+  RadioTower: ["signal", "broadcast", "connectivity", "live", "frequency"],
   Radius: [
     "shape",
     "circle",
@@ -5790,26 +5598,6 @@ export const TAGS: Record<string, string[]> = {
     "size",
   ],
   RailSymbol: ["railway", "train", "track", "line"],
-  Rainbow: [
-    "colors",
-    "colours",
-    "spectrum",
-    "light",
-    "prism",
-    "arc",
-    "clear",
-    "sunshine",
-  ],
-  Rat: [
-    "mouse",
-    "mice",
-    "gerbil",
-    "rodent",
-    "pet",
-    "pest",
-    "plague",
-    "disease",
-  ],
   Ratio: [
     "screens",
     "sizes",
@@ -5830,6 +5618,26 @@ export const TAGS: Record<string, string[]> = {
     "orientation",
     "portrait",
     "landscape",
+  ],
+  Rat: [
+    "mouse",
+    "mice",
+    "gerbil",
+    "rodent",
+    "pet",
+    "pest",
+    "plague",
+    "disease",
+  ],
+  Rainbow: [
+    "colors",
+    "colours",
+    "spectrum",
+    "light",
+    "prism",
+    "arc",
+    "clear",
+    "sunshine",
   ],
   ReceiptCent: [
     "bill",
@@ -5922,6 +5730,14 @@ export const TAGS: Record<string, string[]> = {
     "usd",
     "$",
   ],
+  RectangleHorizontal: [
+    "rectangle",
+    "aspect ratio",
+    "16:9",
+    "horizontal",
+    "shape",
+  ],
+  RectangleVertical: ["rectangle", "aspect ratio", "9:16", "vertical", "shape"],
   RectangleEllipsis: [
     "login",
     "password",
@@ -5946,14 +5762,6 @@ export const TAGS: Record<string, string[]> = {
     "…",
     "...",
   ],
-  RectangleHorizontal: [
-    "rectangle",
-    "aspect ratio",
-    "16:9",
-    "horizontal",
-    "shape",
-  ],
-  RectangleVertical: ["rectangle", "aspect ratio", "9:16", "vertical", "shape"],
   Recycle: ["sustainability", "salvage", "arrows"],
   Redo2: ["undo", "rerun", "history"],
   RedoDot: ["redo", "history", "step", "over", "forward"],
@@ -5997,15 +5805,6 @@ export const TAGS: Record<string, string[]> = {
     "disconnect",
     "ignore",
   ],
-  Refrigerator: [
-    "frigerator",
-    "fridge",
-    "freezer",
-    "cooler",
-    "icebox",
-    "chiller",
-    "cold storage",
-  ],
   RefreshCw: [
     "rotate",
     "reload",
@@ -6015,6 +5814,15 @@ export const TAGS: Record<string, string[]> = {
     "arrows",
     "circular",
     "cycle",
+  ],
+  Refrigerator: [
+    "frigerator",
+    "fridge",
+    "freezer",
+    "cooler",
+    "icebox",
+    "chiller",
+    "cold storage",
   ],
   Regex: ["search", "text", "code"],
   RemoveFormatting: [
@@ -6031,21 +5839,31 @@ export const TAGS: Record<string, string[]> = {
   Repeat1: ["replay"],
   Repeat2: ["arrows", "retweet", "repost", "share", "repeat", "loop"],
   Repeat: ["loop", "arrows"],
-  ReplaceAll: ["search", "substitute", "swap", "change"],
   ReplyAll: ["email"],
-  Reply: ["email"],
   Replace: ["search", "substitute", "swap", "change"],
+  ReplaceAll: ["search", "substitute", "swap", "change"],
+  Reply: ["email"],
   Rewind: ["music"],
-  Ribbon: ["awareness", "strip", "band", "tape", "strap", "cordon"],
   Rocket: ["release", "boost", "launch", "space", "version"],
+  Ribbon: ["awareness", "strip", "band", "tape", "strap", "cordon"],
   RockingChair: ["chair", "furniture", "seat"],
-  Rotate3d: ["gizmo", "transform", "orientation", "orbit", "axis"],
   RollerCoaster: [
     "attraction",
     "entertainment",
     "amusement park",
     "theme park",
     "funfair",
+  ],
+  Rotate3d: ["gizmo", "transform", "orientation", "orbit", "axis"],
+  RotateCcwSquare: [
+    "left",
+    "counter-clockwise",
+    "rotate",
+    "image",
+    "90",
+    "45",
+    "degrees",
+    "°",
   ],
   RotateCcw: [
     "arrow",
@@ -6068,16 +5886,6 @@ export const TAGS: Record<string, string[]> = {
     "degrees",
     "°",
   ],
-  RotateCcwSquare: [
-    "left",
-    "counter-clockwise",
-    "rotate",
-    "image",
-    "90",
-    "45",
-    "degrees",
-    "°",
-  ],
   RotateCw: [
     "arrow",
     "right",
@@ -6087,7 +5895,6 @@ export const TAGS: Record<string, string[]> = {
     "rerun",
     "redo",
   ],
-  Route: ["path", "journey", "planner", "points", "stops", "stations"],
   RouteOff: [
     "path",
     "journey",
@@ -6101,6 +5908,7 @@ export const TAGS: Record<string, string[]> = {
     "closed",
     "blocked",
   ],
+  Route: ["path", "journey", "planner", "points", "stops", "stations"],
   Router: ["computer", "server", "cloud"],
   Rows3: [
     "lines",
@@ -6119,24 +5927,6 @@ export const TAGS: Record<string, string[]> = {
     "even",
     "drawers",
   ],
-  Rows4: [
-    "lines",
-    "list",
-    "queue",
-    "preview",
-    "paragraphs",
-    "parallel",
-    "series",
-    "split",
-    "vertical",
-    "horizontal",
-    "half",
-    "center",
-    "middle",
-    "even",
-    "drawers",
-    "grill",
-  ],
   Rows2: [
     "lines",
     "list",
@@ -6154,6 +5944,24 @@ export const TAGS: Record<string, string[]> = {
     "middle",
     "even",
     "drawer",
+  ],
+  Rows4: [
+    "lines",
+    "list",
+    "queue",
+    "preview",
+    "paragraphs",
+    "parallel",
+    "series",
+    "split",
+    "vertical",
+    "horizontal",
+    "half",
+    "center",
+    "middle",
+    "even",
+    "drawers",
+    "grill",
   ],
   Rss: [
     "feed",
@@ -6178,6 +5986,7 @@ export const TAGS: Record<string, string[]> = {
     "weblog",
     "podcast",
   ],
+  RussianRuble: ["currency", "money", "payment"],
   Ruler: [
     "measurements",
     "centimeters",
@@ -6200,7 +6009,6 @@ export const TAGS: Record<string, string[]> = {
     "stationery",
   ],
   Sailboat: ["ship", "boat", "harbor", "harbour", "dock"],
-  RussianRuble: ["currency", "money", "payment"],
   Salad: [
     "food",
     "vegetarian",
@@ -6215,11 +6023,12 @@ export const TAGS: Record<string, string[]> = {
   Sandwich: ["food", "snack", "dish", "restaurant", "lunch", "meal"],
   SatelliteDish: ["antenna", "receiver", "dish aerial", "saucer"],
   Satellite: ["space station", "orbit", "transmitter"],
-  SaveOff: ["floppy disk", "unsalvageable"],
   SaveAll: ["floppy disks", "copy"],
-  Save: ["floppy disk"],
   Scale3d: ["gizmo", "transform", "size", "axis"],
+  SaveOff: ["floppy disk", "unsalvageable"],
+  Save: ["floppy disk"],
   Scale: ["balance", "legal", "license", "right", "rule", "law"],
+  Scaling: ["scale", "resize", "design"],
   ScanBarcode: [
     "checkout",
     "till",
@@ -6232,7 +6041,6 @@ export const TAGS: Record<string, string[]> = {
     "retail",
     "consumer",
   ],
-  Scaling: ["scale", "resize", "design"],
   ScanEye: [
     "preview",
     "zoom",
@@ -6262,14 +6070,15 @@ export const TAGS: Record<string, string[]> = {
     "login",
     "dashed",
   ],
-  ScanQrCode: [
-    "barcode",
-    "scan",
-    "qrcode",
-    "url",
-    "information",
-    "digital",
-    "scanner",
+  ScanSearch: [
+    "preview",
+    "zoom",
+    "expand",
+    "fullscreen",
+    "gallery",
+    "image",
+    "focus",
+    "lens",
   ],
   ScanLine: [
     "checkout",
@@ -6285,18 +6094,7 @@ export const TAGS: Record<string, string[]> = {
     "qr-code",
     "dashed",
   ],
-  ScanSearch: [
-    "preview",
-    "zoom",
-    "expand",
-    "fullscreen",
-    "gallery",
-    "image",
-    "focus",
-    "lens",
-  ],
   ScanText: ["recognition", "read", "translate", "copy", "lines"],
-  ScatterChart: ["statistics", "analytics", "diagram", "graph"],
   Scan: [
     "qr-code",
     "barcode",
@@ -6316,6 +6114,17 @@ export const TAGS: Record<string, string[]> = {
     "area",
     "square",
     "dashed",
+  ],
+  ScatterChart: ["statistics", "analytics", "diagram", "graph"],
+  ScissorsLineDashed: [
+    "cut here",
+    "along",
+    "snip",
+    "chop",
+    "stationery",
+    "crafts",
+    "instructions",
+    "diagram",
   ],
   School: [
     "building",
@@ -6342,19 +6151,8 @@ export const TAGS: Record<string, string[]> = {
     "assignment",
     "exam",
   ],
-  ScissorsLineDashed: [
-    "cut here",
-    "along",
-    "snip",
-    "chop",
-    "stationery",
-    "crafts",
-    "instructions",
-    "diagram",
-  ],
   Scissors: ["cut", "snip", "chop", "stationery", "crafts"],
   ScreenShareOff: ["desktop", "disconnect", "monitor"],
-  ScreenShare: ["host", "desktop", "monitor"],
   ScrollText: [
     "paper",
     "log",
@@ -6369,6 +6167,7 @@ export const TAGS: Record<string, string[]> = {
     "code",
     "coding",
   ],
+  ScreenShare: ["host", "desktop", "monitor"],
   Scroll: [
     "paper",
     "log",
@@ -6393,6 +6192,16 @@ export const TAGS: Record<string, string[]> = {
     "complete",
     "tick",
   ],
+  SearchX: [
+    "find",
+    "scan",
+    "magnifier",
+    "magnifying glass",
+    "stop",
+    "clear",
+    "cancel",
+    "abort",
+  ],
   SearchCode: [
     "find",
     "scan",
@@ -6413,16 +6222,7 @@ export const TAGS: Record<string, string[]> = {
     "abort",
     "/",
   ],
-  SearchX: [
-    "find",
-    "scan",
-    "magnifier",
-    "magnifying glass",
-    "stop",
-    "clear",
-    "cancel",
-    "abort",
-  ],
+  Search: ["find", "scan", "magnifier", "magnifying glass"],
   Section: [
     "mark",
     "typography",
@@ -6433,7 +6233,6 @@ export const TAGS: Record<string, string[]> = {
     "prose",
     "symbol",
   ],
-  Search: ["find", "scan", "magnifier", "magnifying glass"],
   SendHorizontal: [
     "email",
     "message",
@@ -6463,11 +6262,11 @@ export const TAGS: Record<string, string[]> = {
   ],
   SeparatorHorizontal: ["move", "split"],
   SeparatorVertical: ["move", "split"],
+  ServerCrash: ["cloud", "storage", "problem", "error"],
   ServerCog: ["cloud", "storage", "computing", "cog", "gear"],
   ServerOff: ["cloud", "storage"],
-  ServerCrash: ["cloud", "storage", "problem", "error"],
-  Settings2: ["cog", "edit", "gear", "preferences"],
   Server: ["cloud", "storage"],
+  Settings2: ["cog", "edit", "gear", "preferences"],
   Settings: ["cog", "edit", "gear", "preferences"],
   Shapes: [
     "triangle",
@@ -6481,37 +6280,9 @@ export const TAGS: Record<string, string[]> = {
     "blocks",
     "learning",
   ],
-  Share: ["network", "connections"],
   Share2: ["network", "connections"],
+  Share: ["network", "connections"],
   Sheet: ["spreadsheets", "table", "excel"],
-  Shell: [
-    "beach",
-    "sand",
-    "holiday",
-    "sealife",
-    "fossil",
-    "ammonite",
-    "biology",
-    "ocean",
-    "terminal",
-    "command line",
-    "session",
-    "bash",
-    "zsh",
-    "roll",
-    "wrap",
-    "chewing gum",
-    "bubble gum",
-    "sweet",
-    "sugar",
-    "hosepipe",
-    "carpet",
-    "string",
-    "spiral",
-    "spinner",
-    "hypnotise",
-    "hypnosis",
-  ],
   ShieldAlert: [
     "unshielded",
     "cybersecurity",
@@ -6638,34 +6409,33 @@ export const TAGS: Record<string, string[]> = {
     "inactive",
     "/",
   ],
-  ShieldEllipsis: [
-    "cybersecurity",
-    "securing",
-    "protecting",
-    "guarding",
-    "armoring",
-    "armouring",
-    "defending",
-    "blocking",
-    "preventing",
-    "antivirus",
-    "detecting",
-    "scanning",
-    "finding",
-    "auditing",
-    "admin",
-    "verifying",
-    "crest",
-    "upgrading",
-    "loader",
-    "loading",
-    "throbber",
-    "progress",
-    "dots",
-    "more",
-    "etc",
-    "...",
-    "…",
+  Shell: [
+    "beach",
+    "sand",
+    "holiday",
+    "sealife",
+    "fossil",
+    "ammonite",
+    "biology",
+    "ocean",
+    "terminal",
+    "command line",
+    "session",
+    "bash",
+    "zsh",
+    "roll",
+    "wrap",
+    "chewing gum",
+    "bubble gum",
+    "sweet",
+    "sugar",
+    "hosepipe",
+    "carpet",
+    "string",
+    "spiral",
+    "spinner",
+    "hypnotise",
+    "hypnosis",
   ],
   ShieldCheck: [
     "cybersecurity",
@@ -6722,35 +6492,34 @@ export const TAGS: Record<string, string[]> = {
     "crest",
     "bravery",
   ],
-  ShieldMinus: [
-    "unshield",
+  ShieldEllipsis: [
     "cybersecurity",
-    "unsecure",
-    "unguard",
-    "unblock",
+    "securing",
+    "protecting",
+    "guarding",
+    "armoring",
+    "armouring",
+    "defending",
+    "blocking",
+    "preventing",
     "antivirus",
-    "clean",
-    "clear",
-    "disinfect",
-    "patch",
-    "fix",
-    "stop",
-    "cancel",
-    "remove",
-    "relax",
+    "detecting",
+    "scanning",
+    "finding",
+    "auditing",
     "admin",
+    "verifying",
     "crest",
-    "bravery",
-    "weakened",
-    "damaged",
-    "hit",
-    "unarm",
-    "disable",
-    "deactivate",
-    "decommission",
-    "downgraded",
-    "minimum",
-    "-",
+    "upgrading",
+    "loader",
+    "loading",
+    "throbber",
+    "progress",
+    "dots",
+    "more",
+    "etc",
+    "...",
+    "…",
   ],
   ShieldHalf: [
     "cybersecurity",
@@ -6807,6 +6576,36 @@ export const TAGS: Record<string, string[]> = {
     "army",
     "cadet",
     "scout",
+  ],
+  ShieldMinus: [
+    "unshield",
+    "cybersecurity",
+    "unsecure",
+    "unguard",
+    "unblock",
+    "antivirus",
+    "clean",
+    "clear",
+    "disinfect",
+    "patch",
+    "fix",
+    "stop",
+    "cancel",
+    "remove",
+    "relax",
+    "admin",
+    "crest",
+    "bravery",
+    "weakened",
+    "damaged",
+    "hit",
+    "unarm",
+    "disable",
+    "deactivate",
+    "decommission",
+    "downgraded",
+    "minimum",
+    "-",
   ],
   ShieldOff: [
     "unshielded",
@@ -6898,6 +6697,52 @@ export const TAGS: Record<string, string[]> = {
     "medic",
     "+",
   ],
+  ShieldQuestion: [
+    "unshielded",
+    "cybersecurity",
+    "insecure",
+    "unsecured",
+    "safety",
+    "unsafe",
+    "protection",
+    "unprotected",
+    "guardian",
+    "unguarded",
+    "unarmored",
+    "unarmoured",
+    "defenseless",
+    "defenceless",
+    "undefended",
+    "defender",
+    "threat",
+    "prevention",
+    "unprevented",
+    "antivirus",
+    "vigilance",
+    "vigilant",
+    "detection",
+    "undetected",
+    "scan",
+    "find",
+    "exploit",
+    "vulnerability",
+    "vulnerable",
+    "weakness",
+    "infection",
+    "comprimised",
+    "data leak",
+    "audit",
+    "admin",
+    "verification",
+    "unverified",
+    "uncertified",
+    "uncertain",
+    "unknown",
+    "inactive",
+    "crest",
+    "question mark",
+    "?",
+  ],
   ShieldX: [
     "unshielded",
     "cybersecurity",
@@ -6960,52 +6805,6 @@ export const TAGS: Record<string, string[]> = {
     "eliminated",
     "exterminated",
   ],
-  ShieldQuestion: [
-    "unshielded",
-    "cybersecurity",
-    "insecure",
-    "unsecured",
-    "safety",
-    "unsafe",
-    "protection",
-    "unprotected",
-    "guardian",
-    "unguarded",
-    "unarmored",
-    "unarmoured",
-    "defenseless",
-    "defenceless",
-    "undefended",
-    "defender",
-    "threat",
-    "prevention",
-    "unprevented",
-    "antivirus",
-    "vigilance",
-    "vigilant",
-    "detection",
-    "undetected",
-    "scan",
-    "find",
-    "exploit",
-    "vulnerability",
-    "vulnerable",
-    "weakness",
-    "infection",
-    "comprimised",
-    "data leak",
-    "audit",
-    "admin",
-    "verification",
-    "unverified",
-    "uncertified",
-    "uncertain",
-    "unknown",
-    "inactive",
-    "crest",
-    "question mark",
-    "?",
-  ],
   Shield: [
     "cybersecurity",
     "secure",
@@ -7066,6 +6865,7 @@ export const TAGS: Record<string, string[]> = {
     "navy",
     "trip",
   ],
+  Shirt: ["t-shirt", "shopping", "store", "clothing", "clothes"],
   Ship: [
     "boat",
     "knots",
@@ -7081,7 +6881,6 @@ export const TAGS: Record<string, string[]> = {
     "trip",
   ],
   ShoppingBag: ["ecommerce", "cart", "purchase", "store"],
-  Shirt: ["t-shirt", "shopping", "store", "clothing", "clothes"],
   ShoppingBasket: [
     "cart",
     "e-commerce",
@@ -7103,11 +6902,12 @@ export const TAGS: Record<string, string[]> = {
     "ingredients",
   ],
   Shovel: ["dig", "spade", "treasure"],
-  ShowerHead: ["shower", "bath", "bathroom", "amenities", "services"],
   Shrink: ["scale", "fullscreen"],
+  ShowerHead: ["shower", "bath", "bathroom", "amenities", "services"],
   Shrub: ["forest", "undergrowth", "park", "nature"],
   Shuffle: ["music", "random", "reorder"],
   Sigma: ["sum", "calculate", "formula", "maths", "enumeration", "enumerate"],
+  SignalLow: ["connection", "wireless", "gsm", "phone", "2g", "3g", "4g", "5g"],
   SignalHigh: [
     "connection",
     "wireless",
@@ -7118,7 +6918,6 @@ export const TAGS: Record<string, string[]> = {
     "4g",
     "5g",
   ],
-  SignalLow: ["connection", "wireless", "gsm", "phone", "2g", "3g", "4g", "5g"],
   SignalMedium: [
     "connection",
     "wireless",
@@ -7140,7 +6939,6 @@ export const TAGS: Record<string, string[]> = {
     "5g",
     "lost",
   ],
-  Signal: ["connection", "wireless", "gsm", "phone", "2g", "3g", "4g", "5g"],
   Signature: [
     "text",
     "format",
@@ -7164,6 +6962,8 @@ export const TAGS: Record<string, string[]> = {
     "unique",
   ],
   SignpostBig: ["bidirectional", "left", "right", "east", "west"],
+  Signal: ["connection", "wireless", "gsm", "phone", "2g", "3g", "4g", "5g"],
+  Signpost: ["bidirectional", "left", "right", "east", "west"],
   Siren: [
     "police",
     "ambulance",
@@ -7173,22 +6973,21 @@ export const TAGS: Record<string, string[]> = {
     "alarm",
     "light",
   ],
-  Signpost: ["bidirectional", "left", "right", "east", "west"],
-  SkipForward: ["arrow", "skip", "next", "music"],
   SkipBack: ["arrow", "previous", "music"],
-  Slack: ["logo"],
+  SkipForward: ["arrow", "skip", "next", "music"],
   Skull: ["death", "danger", "bone"],
-  Slash: ["divide", "division", "or", "/"],
+  Slack: ["logo"],
   Slice: ["cutter", "scalpel", "knife"],
+  Slash: ["divide", "division", "or", "/"],
   SlidersHorizontal: ["settings", "filters", "controls"],
-  SlidersVertical: ["settings", "controls"],
+  SmartphoneCharging: ["phone", "cellphone", "device", "power", "screen"],
   SmartphoneNfc: [
     "contactless",
     "payment",
     "near-field communication",
     "screen",
   ],
-  SmartphoneCharging: ["phone", "cellphone", "device", "power", "screen"],
+  SlidersVertical: ["settings", "controls"],
   Smartphone: ["phone", "cellphone", "device", "screen"],
   SmilePlus: [
     "emoji",
@@ -7200,13 +6999,14 @@ export const TAGS: Record<string, string[]> = {
     "reaction",
     "add",
   ],
-  Snail: ["animal", "insect", "slow", "speed", "delicacy", "spiral"],
   Smile: ["emoji", "face", "happy", "good", "emotion"],
+  Snail: ["animal", "insect", "slow", "speed", "delicacy", "spiral"],
   Snowflake: ["cold", "weather", "freeze", "snow", "winter"],
   Sofa: ["armchair", "furniture", "leisure", "lounge", "loveseat", "couch"],
   Soup: ["food", "dish", "restaurant", "course", "meal", "bowl", "starter"],
   Space: ["text", "selection", "letters", "characters", "font", "typography"],
   Spade: ["shape", "suit", "playing", "cards"],
+  Sparkles: ["stars", "effect", "filter", "night", "magic"],
   Sparkle: [
     "star",
     "effect",
@@ -7229,7 +7029,6 @@ export const TAGS: Record<string, string[]> = {
     "producer",
     "dj",
   ],
-  Sparkles: ["stars", "effect", "filter", "night", "magic"],
   Speech: [
     "disability",
     "disabled",
@@ -7250,7 +7049,6 @@ export const TAGS: Record<string, string[]> = {
     "linter",
     "a",
   ],
-  Spline: ["path", "pen", "tool", "shape", "curve", "draw"],
   SpellCheck: [
     "spelling",
     "error",
@@ -7262,6 +7060,7 @@ export const TAGS: Record<string, string[]> = {
     "linter",
     "a",
   ],
+  Spline: ["path", "pen", "tool", "shape", "curve", "draw"],
   Split: ["break", "disband", "divide", "separate", "branch", "disunite"],
   SprayCan: [
     "paint",
@@ -7321,15 +7120,6 @@ export const TAGS: Record<string, string[]> = {
     "keyboard",
     "button",
   ],
-  SquareArrowDown: [
-    "backwards",
-    "reverse",
-    "direction",
-    "south",
-    "sign",
-    "keyboard",
-    "button",
-  ],
   SquareArrowLeft: [
     "previous",
     "back",
@@ -7339,6 +7129,15 @@ export const TAGS: Record<string, string[]> = {
     "keyboard",
     "button",
     "<-",
+  ],
+  SquareArrowDown: [
+    "backwards",
+    "reverse",
+    "direction",
+    "south",
+    "sign",
+    "keyboard",
+    "button",
   ],
   SquareArrowOutDownLeft: ["outwards", "direction", "south-west", "diagonal"],
   SquareArrowOutDownRight: ["outwards", "direction", "south-east", "diagonal"],
@@ -7371,15 +7170,6 @@ export const TAGS: Record<string, string[]> = {
     "keyboard",
     "button",
   ],
-  SquareArrowUpRight: [
-    "direction",
-    "north-east",
-    "diagonal",
-    "sign",
-    "keyboard",
-    "button",
-    "share",
-  ],
   SquareArrowUp: [
     "forward",
     "direction",
@@ -7388,12 +7178,14 @@ export const TAGS: Record<string, string[]> = {
     "keyboard",
     "button",
   ],
-  SquareBottomDashedScissors: [
-    "cut",
-    "snippet",
-    "chop",
-    "stationery",
-    "crafts",
+  SquareArrowUpRight: [
+    "direction",
+    "north-east",
+    "diagonal",
+    "sign",
+    "keyboard",
+    "button",
+    "share",
   ],
   SquareAsterisk: [
     "password",
@@ -7406,6 +7198,15 @@ export const TAGS: Record<string, string[]> = {
     "wildcard",
     "*",
   ],
+  SquareBottomDashedScissors: [
+    "cut",
+    "snippet",
+    "chop",
+    "stationery",
+    "crafts",
+  ],
+  SquareChevronDown: ["back", "menu", "panel"],
+  SquareCheckBig: ["done", "todo", "tick", "complete", "task"],
   SquareCheck: ["done", "todo", "tick", "complete", "task"],
   SquareChevronLeft: [
     "back",
@@ -7418,8 +7219,6 @@ export const TAGS: Record<string, string[]> = {
     "keyboard",
     "<",
   ],
-  SquareCheckBig: ["done", "todo", "tick", "complete", "task"],
-  SquareChevronDown: ["back", "menu", "panel"],
   SquareChevronRight: [
     "forward",
     "next",
@@ -7461,6 +7260,15 @@ export const TAGS: Record<string, string[]> = {
     "code",
     "coding",
   ],
+  SquareDashedBottom: [
+    "rectangle",
+    "aspect ratio",
+    "1:1",
+    "shape",
+    "snippet",
+    "code",
+    "coding",
+  ],
   SquareDashedKanban: [
     "projects",
     "manage",
@@ -7480,15 +7288,6 @@ export const TAGS: Record<string, string[]> = {
     "code",
     "coding",
   ],
-  SquareDashedBottom: [
-    "rectangle",
-    "aspect ratio",
-    "1:1",
-    "shape",
-    "snippet",
-    "code",
-    "coding",
-  ],
   SquareDashedMousePointer: [
     "inspector",
     "element",
@@ -7504,8 +7303,8 @@ export const TAGS: Record<string, string[]> = {
   ],
   SquareEqual: ["calculate", "="],
   SquareDivide: ["calculate", "maths", "÷", "/"],
-  SquareFunction: ["programming", "code", "automation", "maths"],
   SquareDot: ["git", "diff", "modified", "."],
+  SquareFunction: ["programming", "code", "automation", "maths"],
   SquareKanban: [
     "projects",
     "manage",
@@ -7516,6 +7315,27 @@ export const TAGS: Record<string, string[]> = {
     "roadmap",
     "plan",
     "intentions",
+    "productivity",
+    "work",
+    "agile",
+    "code",
+    "coding",
+    "toolbar",
+    "button",
+  ],
+  SquareGanttChart: [
+    "projects",
+    "manage",
+    "overview",
+    "roadmap",
+    "plan",
+    "intentions",
+    "timeline",
+    "deadline",
+    "date",
+    "event",
+    "range",
+    "period",
     "productivity",
     "work",
     "agile",
@@ -7548,27 +7368,6 @@ export const TAGS: Record<string, string[]> = {
     "albums",
     "music",
     "package",
-  ],
-  SquareGanttChart: [
-    "projects",
-    "manage",
-    "overview",
-    "roadmap",
-    "plan",
-    "intentions",
-    "timeline",
-    "deadline",
-    "date",
-    "event",
-    "range",
-    "period",
-    "productivity",
-    "work",
-    "agile",
-    "code",
-    "coding",
-    "toolbar",
-    "button",
   ],
   SquareM: ["metro", "subway", "underground", "track", "line"],
   SquareMinus: [
@@ -7617,7 +7416,6 @@ export const TAGS: Record<string, string[]> = {
     "node",
   ],
   SquareParkingOff: ["parking lot", "car park", "no parking"],
-  SquareParking: ["parking lot", "car park"],
   SquarePen: [
     "pencil",
     "change",
@@ -7634,6 +7432,7 @@ export const TAGS: Record<string, string[]> = {
     "stationery",
     "artist",
   ],
+  SquareParking: ["parking lot", "car park"],
   SquarePercent: [
     "verified",
     "unverified",
@@ -7654,17 +7453,7 @@ export const TAGS: Record<string, string[]> = {
     "geometry",
     "formula",
   ],
-  SquarePilcrow: [
-    "paragraph",
-    "mark",
-    "paraph",
-    "blind",
-    "typography",
-    "type",
-    "text",
-    "prose",
-    "symbol",
-  ],
+  SquarePlay: ["music", "audio", "video", "start", "run"],
   SquarePlus: [
     "add",
     "new",
@@ -7686,7 +7475,17 @@ export const TAGS: Record<string, string[]> = {
     "coding",
     "+",
   ],
-  SquarePlay: ["music", "audio", "video", "start", "run"],
+  SquarePilcrow: [
+    "paragraph",
+    "mark",
+    "paraph",
+    "blind",
+    "typography",
+    "type",
+    "text",
+    "prose",
+    "symbol",
+  ],
   SquarePower: [
     "on",
     "off",
@@ -7719,6 +7518,14 @@ export const TAGS: Record<string, string[]> = {
     "toolbar",
     "button",
   ],
+  SquareSigma: [
+    "sum",
+    "calculate",
+    "formula",
+    "maths",
+    "enumeration",
+    "enumerate",
+  ],
   SquareSlash: [
     "git",
     "diff",
@@ -7729,17 +7536,7 @@ export const TAGS: Record<string, string[]> = {
     "or",
     "/",
   ],
-  SquareSigma: [
-    "sum",
-    "calculate",
-    "formula",
-    "maths",
-    "enumeration",
-    "enumerate",
-  ],
-  SquareSplitVertical: ["split", "divide"],
   SquareSplitHorizontal: ["split", "divide"],
-  SquareTerminal: ["code", "command line", "prompt", "shell"],
   SquareStack: [
     "versions",
     "clone",
@@ -7751,6 +7548,8 @@ export const TAGS: Record<string, string[]> = {
     "backup",
     "history",
   ],
+  SquareTerminal: ["code", "command line", "prompt", "shell"],
+  SquareSplitVertical: ["split", "divide"],
   SquareUserRound: ["person", "account", "contact"],
   SquareUser: ["person", "account", "contact"],
   SquareX: [
@@ -7777,15 +7576,14 @@ export const TAGS: Record<string, string[]> = {
     "storage",
     "stash",
   ],
-  Stamp: ["mark", "print", "clone", "loyalty", "library"],
   StarHalf: ["bookmark", "favorite", "like", "review", "rating"],
-  Star: ["bookmark", "favorite", "like", "review", "rating"],
+  Stamp: ["mark", "print", "clone", "loyalty", "library"],
   StarOff: ["dislike", "unlike", "remove", "unrate"],
+  Star: ["bookmark", "favorite", "like", "review", "rating"],
   StepBack: ["arrow", "previous", "music", "left", "reverse"],
   StepForward: ["arrow", "next", "music", "right", "continue"],
-  Sticker: ["reaction", "emotion", "smile", "happy", "feedback"],
   Stethoscope: ["phonendoscope", "medical", "heart", "lungs", "sound"],
-  Store: ["shop", "supermarket", "stand", "boutique", "building"],
+  Sticker: ["reaction", "emotion", "smile", "happy", "feedback"],
   StickyNote: [
     "post-it",
     "comment",
@@ -7804,8 +7602,9 @@ export const TAGS: Record<string, string[]> = {
     "stationary",
     "office",
   ],
-  StretchHorizontal: ["items", "flex", "justify", "distribute"],
+  Store: ["shop", "supermarket", "stand", "boutique", "building"],
   StretchVertical: ["items", "flex", "justify", "distribute"],
+  StretchHorizontal: ["items", "flex", "justify", "distribute"],
   Strikethrough: ["cross out", "delete", "remove", "format"],
   Subscript: ["text"],
   SunDim: ["brightness", "dim", "low", "brightness low"],
@@ -7822,8 +7621,6 @@ export const TAGS: Record<string, string[]> = {
     "system theme",
     "appearance",
   ],
-  Sun: ["brightness", "weather", "light", "summer"],
-  Sunrise: ["weather", "time", "morning", "day"],
   SunSnow: [
     "weather",
     "air conditioning",
@@ -7832,7 +7629,9 @@ export const TAGS: Record<string, string[]> = {
     "cold",
     "seasons",
   ],
-  Sunset: ["weather", "time", "evening", "night"],
+  Sun: ["brightness", "weather", "light", "summer"],
+  Sunrise: ["weather", "time", "morning", "day"],
+  Superscript: ["text", "exponent"],
   SwatchBook: [
     "colors",
     "colours",
@@ -7850,9 +7649,9 @@ export const TAGS: Record<string, string[]> = {
     "textile",
     "carpet",
   ],
-  Superscript: ["text", "exponent"],
-  SwitchCamera: ["photo", "selfie", "front", "back"],
   SwissFranc: ["currency", "money", "payment"],
+  Sunset: ["weather", "time", "evening", "night"],
+  SwitchCamera: ["photo", "selfie", "front", "back"],
   Sword: ["battle", "challenge", "game", "war", "weapon"],
   Swords: ["battle", "challenge", "game", "war", "weapon"],
   Syringe: [
@@ -7865,17 +7664,8 @@ export const TAGS: Record<string, string[]> = {
     "blood",
   ],
   Table2: ["spreadsheet", "grid"],
-  TableCellsSplit: ["spreadsheet", "grid", "row"],
   TableCellsMerge: ["spreadsheet", "grid", "row"],
-  TableProperties: [
-    "property list",
-    "plist",
-    "spreadsheet",
-    "grid",
-    "dictionary",
-    "object",
-    "hash",
-  ],
+  TableCellsSplit: ["spreadsheet", "grid", "row"],
   TableColumnsSplit: [
     "spreadsheet",
     "grid",
@@ -7885,6 +7675,16 @@ export const TAGS: Record<string, string[]> = {
     "separate",
     "segment",
   ],
+  TableProperties: [
+    "property list",
+    "plist",
+    "spreadsheet",
+    "grid",
+    "dictionary",
+    "object",
+    "hash",
+  ],
+  Table: ["spreadsheet", "grid"],
   TableRowsSplit: [
     "spreadsheet",
     "grid",
@@ -7894,9 +7694,8 @@ export const TAGS: Record<string, string[]> = {
     "separate",
     "segment",
   ],
-  Table: ["spreadsheet", "grid"],
-  TabletSmartphone: ["responsive", "screens", "browser", "testing", "mobile"],
   Tablet: ["device"],
+  TabletSmartphone: ["responsive", "screens", "browser", "testing", "mobile"],
   Tablets: [
     "medicine",
     "medication",
@@ -8030,7 +7829,7 @@ export const TAGS: Record<string, string[]> = {
     "wilderness",
     "outdoors",
   ],
-  TestTube: [
+  TestTubeDiagonal: [
     "tube",
     "vial",
     "phial",
@@ -8043,7 +7842,7 @@ export const TAGS: Record<string, string[]> = {
     "test",
   ],
   Terminal: ["code", "command line", "prompt", "shell"],
-  TestTubeDiagonal: [
+  TestTube: [
     "tube",
     "vial",
     "phial",
@@ -8081,8 +7880,8 @@ export const TAGS: Record<string, string[]> = {
     "magnifier",
     "magnifying glass",
   ],
-  Text: ["find", "search", "data", "txt", "pdf", "document"],
   TextSelect: ["find", "search", "selection", "dashed"],
+  Text: ["find", "search", "data", "txt", "pdf", "document"],
   Theater: [
     "theater",
     "theatre",
@@ -8100,8 +7899,6 @@ export const TAGS: Record<string, string[]> = {
     "freeze",
     "freezing",
   ],
-  Thermometer: ["temperature", "celsius", "fahrenheit", "weather"],
-  ThumbsDown: ["dislike", "bad", "emotion"],
   ThermometerSun: [
     "temperature",
     "celsius",
@@ -8110,6 +7907,8 @@ export const TAGS: Record<string, string[]> = {
     "warm",
     "hot",
   ],
+  Thermometer: ["temperature", "celsius", "fahrenheit", "weather"],
+  ThumbsDown: ["dislike", "bad", "emotion"],
   ThumbsUp: ["like", "good", "emotion"],
   TicketCheck: [
     "entry",
@@ -8235,11 +8034,9 @@ export const TAGS: Record<string, string[]> = {
   ],
   TimerOff: ["time", "timer", "stopwatch"],
   TimerReset: ["time", "timer", "stopwatch"],
-  ToggleLeft: ["on", "off", "switch", "boolean"],
   Timer: ["time", "timer", "stopwatch"],
+  ToggleLeft: ["on", "off", "switch", "boolean"],
   ToggleRight: ["on", "off", "switch", "boolean"],
-  Tornado: ["weather", "wind", "storm", "hurricane"],
-  TouchpadOff: ["trackpad", "cursor"],
   Torus: [
     "donut",
     "doughnut",
@@ -8254,10 +8051,13 @@ export const TAGS: Record<string, string[]> = {
     "sugar",
     "dessert",
   ],
+  Tornado: ["weather", "wind", "storm", "hurricane"],
+  TouchpadOff: ["trackpad", "cursor"],
   Touchpad: ["trackpad", "cursor"],
+  Tractor: ["farming", "farmer", "ranch", "harvest", "equipment", "vehicle"],
   TowerControl: ["airport", "travel", "tower", "transportation", "lighthouse"],
   ToyBrick: ["lego", "block", "addon", "plugin", "integration"],
-  Tractor: ["farming", "farmer", "ranch", "harvest", "equipment", "vehicle"],
+  TrafficCone: ["roadworks", "tarmac", "safety", "block"],
   TrainFrontTunnel: [
     "railway",
     "metro",
@@ -8269,7 +8069,6 @@ export const TAGS: Record<string, string[]> = {
     "track",
     "line",
   ],
-  TrafficCone: ["roadworks", "tarmac", "safety", "block"],
   TrainFront: [
     "railway",
     "metro",
@@ -8282,6 +8081,7 @@ export const TAGS: Record<string, string[]> = {
     "line",
   ],
   TrainTrack: ["railway", "line"],
+  Trash2: ["garbage", "delete", "remove", "bin"],
   TramFront: [
     "railway",
     "metro",
@@ -8291,16 +8091,16 @@ export const TAGS: Record<string, string[]> = {
     "line",
     "tourism",
   ],
-  Trash2: ["garbage", "delete", "remove", "bin"],
-  Trash: ["garbage", "delete", "remove", "bin"],
   TreeDeciduous: ["tree", "forest", "park", "nature"],
-  TreePine: ["tree", "pine", "forest", "park", "nature"],
+  Trash: ["garbage", "delete", "remove", "bin"],
   TreePalm: ["vacation", "leisure", "island"],
   Trees: ["tree", "forest", "park", "nature"],
-  Trello: ["logo", "brand"],
+  TreePine: ["tree", "pine", "forest", "park", "nature"],
   TrendingDown: ["statistics"],
-  TrendingUp: ["statistics"],
+  Trello: ["logo", "brand"],
   TriangleAlert: ["warning", "alert", "danger", "exclamation mark", "linter"],
+  Triangle: ["equilateral", "delta", "shape", "pyramid", "hierarchy"],
+  TrendingUp: ["statistics"],
   TriangleRight: [
     "volume",
     "controls",
@@ -8313,9 +8113,8 @@ export const TAGS: Record<string, string[]> = {
     "incline",
     "increase",
   ],
-  Trophy: ["prize", "sports", "winner", "achievement", "award"],
-  Triangle: ["equilateral", "delta", "shape", "pyramid", "hierarchy"],
   Truck: ["delivery", "van", "shipping", "haulage", "lorry"],
+  Trophy: ["prize", "sports", "winner", "achievement", "award"],
   Turtle: ["animal", "pet", "tortoise", "slow", "speed"],
   TvMinimalPlay: [
     "flatscreen",
@@ -8337,26 +8136,6 @@ export const TAGS: Record<string, string[]> = {
     "running",
     "start",
     "film",
-    "home cinema",
-    "entertainment",
-    "showtime",
-    "channels",
-    "catchup",
-  ],
-  TvMinimal: [
-    "flatscreen",
-    "television",
-    "stream",
-    "display",
-    "widescreen",
-    "high-definition",
-    "hd",
-    "1080p",
-    "4k",
-    "8k",
-    "smart",
-    "digital",
-    "video",
     "home cinema",
     "entertainment",
     "showtime",
@@ -8394,18 +8173,28 @@ export const TAGS: Record<string, string[]> = {
     "connection",
     "connectivity",
   ],
+  TvMinimal: [
+    "flatscreen",
+    "television",
+    "stream",
+    "display",
+    "widescreen",
+    "high-definition",
+    "hd",
+    "1080p",
+    "4k",
+    "8k",
+    "smart",
+    "digital",
+    "video",
+    "home cinema",
+    "entertainment",
+    "showtime",
+    "channels",
+    "catchup",
+  ],
   Twitch: ["logo", "social"],
   Twitter: ["logo", "social"],
-  TypeOutline: [
-    "text",
-    "font",
-    "typography",
-    "silhouette",
-    "profile",
-    "contour",
-    "stroke",
-    "line",
-  ],
   UmbrellaOff: [
     "rain",
     "weather",
@@ -8430,8 +8219,9 @@ export const TAGS: Record<string, string[]> = {
     "u-turn",
   ],
   UndoDot: ["redo", "history", "step", "back"],
-  Undo: ["redo", "rerun", "history"],
   UnfoldHorizontal: ["arrow", "collapse", "fold", "vertical", "dashed"],
+  Undo: ["redo", "rerun", "history"],
+  UnfoldVertical: ["arrow", "expand", "vertical", "dashed"],
   Ungroup: [
     "cubes",
     "packages",
@@ -8441,7 +8231,6 @@ export const TAGS: Record<string, string[]> = {
     "cluster",
     "separate",
   ],
-  UnfoldVertical: ["arrow", "expand", "vertical", "dashed"],
   University: ["building", "education", "childhood", "school"],
   Unlink2: ["url", "unchain"],
   Unplug: [
@@ -8453,8 +8242,9 @@ export const TAGS: Record<string, string[]> = {
     "disconnect",
   ],
   Unlink: ["url", "unchain"],
-  Usb: ["universal", "serial", "bus", "controller", "connector", "interface"],
   Upload: ["file"],
+  Usb: ["universal", "serial", "bus", "controller", "connector", "interface"],
+  UserCog: ["settings", "edit", "cog", "gear"],
   UserCheck: [
     "followed",
     "subscribed",
@@ -8464,9 +8254,7 @@ export const TAGS: Record<string, string[]> = {
     "complete",
     "task",
   ],
-  UserCog: ["settings", "edit", "cog", "gear"],
   UserMinus: ["delete", "remove", "unfollow", "unsubscribe"],
-  UserPen: ["person", "account", "contact", "profile", "edit", "change"],
   UserPlus: ["new", "add", "create", "follow", "subscribe"],
   UserRoundCheck: [
     "followed",
@@ -8479,8 +8267,8 @@ export const TAGS: Record<string, string[]> = {
   ],
   UserRoundCog: ["settings", "edit", "cog", "gear"],
   UserRoundMinus: ["delete", "remove", "unfollow", "unsubscribe"],
-  UserRoundPen: ["person", "account", "contact", "profile", "edit", "change"],
   UserRoundPlus: ["new", "add", "create", "follow", "subscribe"],
+  UserRoundX: ["delete", "remove", "unfollow", "unsubscribe", "unavailable"],
   UserRoundSearch: [
     "person",
     "account",
@@ -8490,7 +8278,7 @@ export const TAGS: Record<string, string[]> = {
     "magnifier",
     "magnifying glass",
   ],
-  UserRoundX: ["delete", "remove", "unfollow", "unsubscribe", "unavailable"],
+  UserRound: ["person", "account", "contact"],
   UserSearch: [
     "person",
     "account",
@@ -8500,12 +8288,11 @@ export const TAGS: Record<string, string[]> = {
     "magnifier",
     "magnifying glass",
   ],
-  UserRound: ["person", "account", "contact"],
   UserX: ["delete", "remove", "unfollow", "unsubscribe", "unavailable"],
   User: ["person", "account", "contact"],
   UsersRound: ["group", "people"],
   Users: ["group", "people"],
-  Utensils: [
+  UtensilsCrossed: [
     "food",
     "restaurant",
     "meal",
@@ -8514,7 +8301,7 @@ export const TAGS: Record<string, string[]> = {
     "dinner",
     "supper",
   ],
-  UtensilsCrossed: [
+  Utensils: [
     "food",
     "restaurant",
     "meal",
@@ -8559,7 +8346,7 @@ export const TAGS: Record<string, string[]> = {
     "valuables",
     "bank",
   ],
-  VenetianMask: ["mask", "masquerade", "impersonate", "secret", "incognito"],
+  Vegan: ["vegetarian", "fruitarian", "herbivorous", "animal rights", "diet"],
   VibrateOff: [
     "smartphone",
     "notification",
@@ -8568,8 +8355,8 @@ export const TAGS: Record<string, string[]> = {
     "notifications",
     "screen",
   ],
+  VenetianMask: ["mask", "masquerade", "impersonate", "secret", "incognito"],
   VideoOff: ["camera", "movie", "film"],
-  Vegan: ["vegetarian", "fruitarian", "herbivorous", "animal rights", "diet"],
   Vibrate: [
     "smartphone",
     "notification",
@@ -8596,11 +8383,10 @@ export const TAGS: Record<string, string[]> = {
     "cassette",
   ],
   Voicemail: ["phone", "cassette", "tape", "reel", "recording", "audio"],
+  View: ["eye", "look"],
   Volume1: ["music", "sound", "speaker"],
   Volume2: ["music", "sound", "speaker"],
-  View: ["eye", "look"],
   VolumeX: ["music", "sound", "mute", "speaker"],
-  Vote: ["vote", "poll", "ballot", "political", "social", "check", "tick"],
   Volume: ["music", "sound", "mute", "speaker"],
   WalletCards: [
     "money",
@@ -8614,25 +8400,15 @@ export const TAGS: Record<string, string[]> = {
     "consumer",
     "cc",
   ],
+  Vote: ["vote", "poll", "ballot", "political", "social", "check", "tick"],
   WalletMinimal: ["finance", "pocket"],
   Wallet: ["money", "finance", "pocket"],
-  Wallpaper: ["cover", "lock screen"],
   WandSparkles: ["magic", "wizard", "magician"],
+  Wallpaper: ["cover", "lock screen"],
   Wand: ["magic", "selection"],
-  WashingMachine: [
-    "tumble dryer",
-    "amenities",
-    "electronics",
-    "cycle",
-    "clothes",
-    "rinse",
-    "spin",
-    "drum",
-  ],
-  Watch: ["clock", "time"],
   Warehouse: ["storage", "logistics", "building"],
+  Watch: ["clock", "time"],
   Waves: ["water", "sea", "sound", "hertz", "wavelength", "vibrate"],
-  WebhookOff: ["push api", "interface", "callback"],
   Waypoints: [
     "indirection",
     "vpn",
@@ -8650,7 +8426,18 @@ export const TAGS: Record<string, string[]> = {
     "spread",
     "viral",
   ],
+  WashingMachine: [
+    "tumble dryer",
+    "amenities",
+    "electronics",
+    "cycle",
+    "clothes",
+    "rinse",
+    "spin",
+    "drum",
+  ],
   Webcam: ["camera", "security"],
+  Webhook: ["push api", "interface", "callback"],
   Weight: [
     "mass",
     "heavy",
@@ -8661,7 +8448,7 @@ export const TAGS: Record<string, string[]> = {
     "scales",
     "balance",
   ],
-  Webhook: ["push api", "interface", "callback"],
+  WebhookOff: ["push api", "interface", "callback"],
   WheatOff: [
     "corn",
     "cereal",
@@ -8680,11 +8467,8 @@ export const TAGS: Record<string, string[]> = {
     "font",
     "typography",
   ],
-  WifiHigh: ["connection", "signal", "wireless"],
   WifiOff: ["disabled"],
-  WifiLow: ["connection", "signal", "wireless"],
   Wifi: ["connection", "signal", "wireless"],
-  WifiZero: ["connection", "signal", "wireless"],
   Wind: ["weather", "air", "blow"],
   WineOff: [
     "alcohol",
@@ -8730,6 +8514,16 @@ export const TAGS: Record<string, string[]> = {
     "computer virus",
     "malware",
   ],
+  WrapText: ["words", "lines", "break", "paragraph"],
+  Wrench: [
+    "account",
+    "settings",
+    "spanner",
+    "diy",
+    "toolbox",
+    "build",
+    "construction",
+  ],
   X: [
     "cancel",
     "close",
@@ -8740,16 +8534,6 @@ export const TAGS: Record<string, string[]> = {
     "maths",
     "multiply",
     "multiplication",
-  ],
-  WrapText: ["words", "lines", "break", "paragraph"],
-  Wrench: [
-    "account",
-    "settings",
-    "spanner",
-    "diy",
-    "toolbox",
-    "build",
-    "construction",
   ],
   Youtube: ["logo", "social", "video", "play"],
   ZapOff: ["flash", "camera", "lightning", "electricity", "energy"],
