+++ docs/pages/index.page.tsx
@@ -13,9 +13,8 @@ import { Section } from "components/layouts"
 import { SEO, OpenCollective, Patreon } from "components/media-and-icons"
 import { NextLinkButton } from "components/navigation"
 import { CONSTANT } from "constant"
-import { useI18n } from "contexts/i18n-context"
-import { PageProvider } from "contexts/page-context"
-import { TopLayout } from "layouts/top-layout"
+import { useI18n, PageProvider } from "contexts"
+import { TopLayout } from "layouts"
 import type { InferGetStaticPropsType, NextPage } from "next"
 import { getStaticCommonProps } from "utils/next"
