+++ docs/scripts/icon-tags.ts
@@ -1,10 +1,10 @@
+import { readdir, readFile, writeFile } from "fs/promises"
+import path from "path"
 import { config } from "dotenv"
 import { execa } from "execa"
-import { rimraf } from "rimraf"
-import path from "path"
-import { readdir, readFile, writeFile } from "fs/promises"
-import { toCamelCase } from "utils/string"
 import { prettier } from "libs/prettier"
+import { rimraf } from "rimraf"
+import { toCamelCase } from "utils/string"
 
 const DIR_PATH = path.resolve(".lucide")
