+++ docs/utils/github.ts
@@ -1,12 +1,12 @@
+import crypto from "crypto"
+import type { RequestError } from "@octokit/request-error"
+import { Octokit } from "@octokit/rest"
 import type {
   EmitterWebhookEvent,
   EmitterWebhookEventName,
 } from "@octokit/webhooks"
 import type { NextApiRequest } from "next"
-import { Octokit } from "@octokit/rest"
-import type { RequestError } from "@octokit/request-error"
 import { wait } from "./async"
-import crypto from "crypto"
 
 const COMMON_PARAMS = {
   owner: "yamada-ui",
